Network.AWS.DynamoDB

Description

Amazon DynamoDB is a fully managed NoSQL database service that provides fast and predictable performance with seamless scalability. You can use Amazon DynamoDB to create a database table that can store and retrieve any amount of data, and serve any level of request traffic. Amazon DynamoDB automatically spreads the data and traffic for the table over a sufficient number of servers to handle the request capacity specified by the customer and the amount of data stored, while maintaining consistent and fast performance.

Documentation

module Network.AWS.DynamoDB.BatchGetItem

module Network.AWS.DynamoDB.BatchWriteItem

module Network.AWS.DynamoDB.CreateTable

module Network.AWS.DynamoDB.DeleteItem

module Network.AWS.DynamoDB.DeleteTable

module Network.AWS.DynamoDB.DescribeTable

module Network.AWS.DynamoDB.GetItem

module Network.AWS.DynamoDB.ListTables

module Network.AWS.DynamoDB.PutItem

module Network.AWS.DynamoDB.Query

module Network.AWS.DynamoDB.Scan

module Network.AWS.DynamoDB.Types

module Network.AWS.DynamoDB.UpdateItem

module Network.AWS.DynamoDB.UpdateTable

module Network.AWS.DynamoDB.Waiters