Network.AWS.DynamoDB.DeleteTable

Description

The DeleteTable operation deletes a table and all of its items. After a DeleteTable request, the specified table is in the DELETING state until DynamoDB completes the deletion. If the table is in the ACTIVE state, you can delete it. If a table is in CREATING or UPDATING states, then DynamoDB returns a ResourceInUseException. If the specified table does not exist, DynamoDB returns a ResourceNotFoundException. If table is already in the DELETING state, no error is returned.

DynamoDB might continue to accept data read and write operations, such as GetItem and PutItem, on a table in the DELETING state until the table deletion is complete.

When you delete a table, any indexes on that table are also deleted.

Use the DescribeTable API to check the status of the table.

http://docs.aws.amazon.com/amazondynamodb/latest/APIReference/API_DeleteTable.html

Request

data DeleteTable

Request constructor

deleteTable

Arguments

:: Text
dtTableName

-> DeleteTable

DeleteTable constructor.

The fields accessible through corresponding lenses are:

dtTableName :: Text

Request lenses

dtTableName :: Lens' DeleteTable Text

The name of the table to delete.

Response

data DeleteTableResponse

Response constructor

deleteTableResponse :: DeleteTableResponse

DeleteTableResponse constructor.

The fields accessible through corresponding lenses are:

dtrTableDescription :: Maybe TableDescription

Response lenses

dtrTableDescription :: Lens' DeleteTableResponse (Maybe TableDescription)