Network.AWS.DynamoDB.DescribeTable

Description

Returns information about the table, including the current status of the table, when it was created, the primary key schema, and any indexes on the table. Describe a Table This example describes the Thread table. { Table: { AttributeDefinitions: [ { AttributeName: ForumName, AttributeType: S }, { AttributeName: LastPostDateTime, AttributeType: S }, { AttributeName: Subject, AttributeType: S } ], CreationDateTime: 1.363729002358E9, ItemCount: 0, KeySchema: [ { AttributeName: ForumName, KeyType: HASH }, { AttributeName: Subject, KeyType: RANGE } ], LocalSecondaryIndexes: [ { IndexName: LastPostIndex, IndexSizeBytes: 0, ItemCount: 0, KeySchema: [ { AttributeName: ForumName, KeyType: HASH }, { AttributeName: LastPostDateTime, KeyType: RANGE } ], Projection: { ProjectionType: KEYS_ONLY } } ], ProvisionedThroughput: { NumberOfDecreasesToday: 0, ReadCapacityUnits: 5, WriteCapacityUnits: 5 }, TableName: Thread, TableSizeBytes: 0, TableStatus: ACTIVE } }.

Request

data DescribeTable

Represents the input of a DescribeTable operation.

Request constructor

describeTable

Arguments

:: Text
dt1TableName

-> DescribeTable

Smart constructor for the minimum required parameters to construct a valid DescribeTable request.

The fields accessible through corresponding lenses are:

TableName :: Text

Request lenses

dt1TableName :: Lens' DescribeTable Text

The name of the table to describe.

Response

data DescribeTableResponse

Represents the output of a DescribeTable operation.

Response constructor

describeTableResponse :: DescribeTableResponse

Smart constructor for the minimum required parameters to construct a valid DescribeTableResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Table :: Maybe TableDescription

Response lenses

dtrrTable :: Lens' DescribeTableResponse (Maybe TableDescription)

Represents the properties of a table.