Network.AWS.DynamoDB.DescribeTimeToLive

Description

Gives a description of the Time to Live (TTL) status on the specified table.

Creating a Request

describeTimeToLive

Arguments

:: Text
dttlTableName

-> DescribeTimeToLive

Creates a value of DescribeTimeToLive with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dttlTableName - The name of the table to be described.

data DescribeTimeToLive

See: describeTimeToLive smart constructor.

Request Lenses

dttlTableName :: Lens' DescribeTimeToLive Text

The name of the table to be described.

Destructuring the Response

describeTimeToLiveResponse

Arguments

:: Int
dttlrsResponseStatus

-> DescribeTimeToLiveResponse

Creates a value of DescribeTimeToLiveResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dttlrsTimeToLiveDescription -

dttlrsResponseStatus - -- | The response status code.

data DescribeTimeToLiveResponse

See: describeTimeToLiveResponse smart constructor.

Response Lenses

dttlrsTimeToLiveDescription :: Lens' DescribeTimeToLiveResponse (Maybe TimeToLiveDescription)

dttlrsResponseStatus :: Lens' DescribeTimeToLiveResponse Int

- | The response status code.