Network.AWS.DynamoDB.ListTables

Description

Returns an array of table names associated with the current account and endpoint. The output from ListTables is paginated, with each page returning a maximum of 100 table names.

http://docs.aws.amazon.com/amazondynamodb/latest/APIReference/API_ListTables.html

Request

data ListTables

Request constructor

listTables :: ListTables

ListTables constructor.

The fields accessible through corresponding lenses are:

ltExclusiveStartTableName :: Maybe Text

ltLimit :: Maybe Natural

Request lenses

ltExclusiveStartTableName :: Lens' ListTables (Maybe Text)

The first table name that this operation will evaluate. Use the value that was returned for LastEvaluatedTableName in a previous operation, so that you can obtain the next page of results.

ltLimit :: Lens' ListTables (Maybe Natural)

A maximum number of table names to return. If this parameter is not specified, the limit is 100.

Response

data ListTablesResponse

Response constructor

listTablesResponse :: ListTablesResponse

ListTablesResponse constructor.

The fields accessible through corresponding lenses are:

ltrLastEvaluatedTableName :: Maybe Text

ltrTableNames :: [Text]

Response lenses

ltrLastEvaluatedTableName :: Lens' ListTablesResponse (Maybe Text)

The name of the last table in the current page of results. Use this value as the ExclusiveStartTableName in a new request to obtain the next page of results, until all the table names are returned.

If you do not receive a LastEvaluatedTableName value in the response, this means that there are no more table names to be retrieved.

ltrTableNames :: Lens' ListTablesResponse [Text]

The names of the tables associated with the current account at the current endpoint. The maximum size of this array is 100.

If LastEvaluatedTableName also appears in the output, you can use this value as the ExclusiveStartTableName parameter in a subsequent ListTables request and obtain the next page of results.