Network.AWS.DynamoDB.Scan

Description

The Scan operation returns one or more items and item attributes by accessing every item in the table. To have DynamoDB return fewer items, you can provide a ScanFilter operation. If the total number of scanned items exceeds the maximum data set size limit of 1 MB, the scan stops and results are returned to the user as a LastEvaluatedKey value to continue the scan in a subsequent operation. The results also include the number of items exceeding the limit. A scan can result in no table data meeting the filter criteria. The result set is eventually consistent. By default, Scan operations proceed sequentially; however, for faster performance on large tables, applications can request a parallel Scan operation by specifying the Segment and TotalSegments parameters. For more information, see Parallel Scan in the Amazon DynamoDB Developer Guide.

http://docs.aws.amazon.com/amazondynamodb/latest/APIReference/API_Scan.html

Request

data Scan

Request constructor

scan

Arguments

:: Text
sTableName

-> NonEmpty Text
sAttributesToGet

-> Scan

Scan constructor.

The fields accessible through corresponding lenses are:

sAttributesToGet :: NonEmpty Text

sConditionalOperator :: Maybe ConditionalOperator

sExclusiveStartKey :: HashMap Text AttributeValue

sExpressionAttributeNames :: HashMap Text Text

sExpressionAttributeValues :: HashMap Text AttributeValue

sFilterExpression :: Maybe Text

sLimit :: Maybe Natural

sProjectionExpression :: Maybe Text

sReturnConsumedCapacity :: Maybe ReturnConsumedCapacity

sScanFilter :: HashMap Text Condition

sSegment :: Maybe Natural

sSelect :: Maybe Select

sTableName :: Text

sTotalSegments :: Maybe Natural

Request lenses

sAttributesToGet :: Lens' Scan (NonEmpty Text)

There is a newer parameter available. Use ProjectionExpression instead. Note that if you use AttributesToGet and ProjectionExpression at the same time, DynamoDB will return a ValidationException exception. This parameter allows you to retrieve lists or maps; however, it cannot retrieve individual list or map elements. The names of one or more attributes to retrieve. If no attribute names are specified, then all attributes will be returned. If any of the requested attributes are not found, they will not appear in the result. Note that AttributesToGet has no effect on provisioned throughput consumption. DynamoDB determines capacity units consumed based on item size, not on the amount of data that is returned to an application.

sConditionalOperator :: Lens' Scan (Maybe ConditionalOperator)

There is a newer parameter available. Use ConditionExpression instead. Note that if you use ConditionalOperator and ConditionExpression at the same time, DynamoDB will return a ValidationException exception. This parameter does not support lists or maps. A logical operator to apply to the conditions in the ScanFilter map: AND - If all of the conditions evaluate to true, then the entire map evaluates to true. OR - If at least one of the conditions evaluate to true, then the entire map evaluates to true. If you omit ConditionalOperator, then AND is the default. The operation will succeed only if the entire map evaluates to true.

sExclusiveStartKey :: Lens' Scan (HashMap Text AttributeValue)

The primary key of the first item that this operation will evaluate. Use the value that was returned for LastEvaluatedKey in the previous operation. The data type for ExclusiveStartKey must be String, Number or Binary. No set data types are allowed. In a parallel scan, a Scan request that includes ExclusiveStartKey must specify the same segment whose previous Scan returned the corresponding value of LastEvaluatedKey.

sExpressionAttributeNames :: Lens' Scan (HashMap Text Text)

One or more substitution tokens for simplifying complex expressions. The following are some use cases for an ExpressionAttributeNames value: To shorten an attribute name that is very long or unwieldy in an expression. To create a placeholder for repeating occurrences of an attribute name in an expression. To prevent special characters in an attribute name from being misinterpreted in an expression. Use the # character in an expression to dereference an attribute name. For example, consider the following expression: order.customerInfo.LastName = Smith OR order.customerInfo.LastName = Jones Now suppose that you specified the following for ExpressionAttributeNames: {"n":"order.customerInfo.LastName"} The expression can now be simplified as follows: n = Jones.

sExpressionAttributeValues :: Lens' Scan (HashMap Text AttributeValue)

One or more values that can be substituted in an expression. Use the : character in an expression to dereference an attribute value. For example, consider the following expression: ProductStatus IN (Available,Backordered,Discontinued) Now suppose that you specified the following for ExpressionAttributeValues: { "a":{S:Available}, "b":{S:Backordered}, "d":{S:Discontinued} } The expression can now be simplified as follows: ProductStatus IN (:a,:b,:c).

sFilterExpression :: Lens' Scan (Maybe Text)

A condition that evaluates the scan results and returns only the desired values. The condition you specify is applied to the items scanned; any items that do not match the expression are not returned.

sLimit :: Lens' Scan (Maybe Natural)

The maximum number of items to evaluate (not necessarily the number of matching items). If DynamoDB processes the number of items up to the limit while processing the results, it stops the operation and returns the matching values up to that point, and a key in LastEvaluatedKey to apply in a subsequent operation, so that you can pick up where you left off. Also, if the processed data set size exceeds 1 MB before DynamoDB reaches this limit, it stops the operation and returns the matching values up to the limit, and a key in LastEvaluatedKey to apply in a subsequent operation to continue the operation. For more information, see Query and Scan in the Amazon DynamoDB Developer Guide.

sProjectionExpression :: Lens' Scan (Maybe Text)

One or more attributes to retrieve from the table. These attributes can include scalars, sets, or elements of a JSON document. The attributes in the expression must be separated by commas. If no attribute names are specified, then all attributes will be returned. If any of the requested attributes are not found, they will not appear in the result.

sReturnConsumedCapacity :: Lens' Scan (Maybe ReturnConsumedCapacity)

sScanFilter :: Lens' Scan (HashMap Text Condition)

There is a newer parameter available. Use FilterExpression instead. Note that if you use ScanFilter and FilterExpression at the same time, DynamoDB will return a ValidationException exception. This parameter does not support lists or maps. A condition that evaluates the scan results and returns only the desired values. If you specify more than one condition in the ScanFilter map, then by default all of the conditions must evaluate to true. In other words, the conditions are ANDed together. (You can use the ConditionalOperator parameter to OR the conditions instead. If you do this, then at least one of the conditions must evaluate to true, rather than all of them.) Each ScanFilter element consists of an attribute name to compare, along with the following: AttributeValueList - One or more values to evaluate against the supplied attribute. The number of values in the list depends on the operator specified in ComparisonOperator . For type Number, value comparisons are numeric. String value comparisons for greater than, equals, or less than are based on ASCII character code values. For example, a is greater than A, and aa is greater than B. For a list of code values, see http:/en.wikipedia.orgwiki/ASCII#ASCII_printable_characters. For Binary, DynamoDB treats each byte of the binary data as unsigned when it compares binary values, for example when evaluating query expressions. For information on specifying data types in JSON, see JSON Data Format in the Amazon DynamoDB Developer Guide. ComparisonOperator - A comparator for evaluating attributes. For example, equals, greater than, less than, etc. The following comparison operators are available: EQ | NE | LE | LT | GE | GT | NOT_NULL | NULL | CONTAINS | NOT_CONTAINS | BEGINS_WITH | IN | BETWEEN For complete descriptions of all comparison operators, see Condition.

sSegment :: Lens' Scan (Maybe Natural)

For a parallel Scan request, Segment identifies an individual segment to be scanned by an application worker. Segment IDs are zero-based, so the first segment is always 0. For example, if you want to scan a table using four application threads, the first thread specifies a Segment value of 0, the second thread specifies 1, and so on. The value of LastEvaluatedKey returned from a parallel Scan request must be used as ExclusiveStartKey with the same segment ID in a subsequent Scan operation. The value for Segment must be greater than or equal to 0, and less than the value provided for TotalSegments. If you specify Segment, you must also specify TotalSegments.

sSelect :: Lens' Scan (Maybe Select)

The attributes to be returned in the result. You can retrieve all item attributes, specific item attributes, or the count of matching items. ALL_ATTRIBUTES - Returns all of the item attributes. COUNT - Returns the number of matching items, rather than the matching items themselves. SPECIFIC_ATTRIBUTES - Returns only the attributes listed in AttributesToGet. This return value is equivalent to specifying AttributesToGet without specifying any value for Select. If neither Select nor AttributesToGet are specified, DynamoDB defaults to ALL_ATTRIBUTES. You cannot use both AttributesToGet and Select together in a single request, unless the value for Select is SPECIFIC_ATTRIBUTES. (This usage is equivalent to specifying AttributesToGet without any value for Select.).

sTableName :: Lens' Scan Text

The name of the table containing the requested items.

sTotalSegments :: Lens' Scan (Maybe Natural)

For a parallel Scan request, TotalSegments represents the total number of segments into which the Scan operation will be divided. The value of TotalSegments corresponds to the number of application workers that will perform the parallel scan. For example, if you want to scan a table using four application threads, specify a TotalSegments value of 4. The value for TotalSegments must be greater than or equal to 1, and less than or equal to 1000000. If you specify a TotalSegments value of 1, the Scan operation will be sequential rather than parallel. If you specify TotalSegments, you must also specify Segment.

Response

data ScanResponse

Response constructor

scanResponse :: ScanResponse

ScanResponse constructor.

The fields accessible through corresponding lenses are:

srConsumedCapacity :: Maybe ConsumedCapacity

srCount :: Maybe Int

srItems :: [HashMap Text AttributeValue]

srLastEvaluatedKey :: HashMap Text AttributeValue

srScannedCount :: Maybe Int

Response lenses

srConsumedCapacity :: Lens' ScanResponse (Maybe ConsumedCapacity)

srCount :: Lens' ScanResponse (Maybe Int)

The number of items in the response. If you set ScanFilter in the request, then Count is the number of items returned after the filter was applied, and ScannedCount is the number of matching items before the filter was applied. If you did not use a filter in the request, then Count is the same as ScannedCount.

srItems :: Lens' ScanResponse [HashMap Text AttributeValue]

An array of item attributes that match the scan criteria. Each element in this array consists of an attribute name and the value for that attribute.

srLastEvaluatedKey :: Lens' ScanResponse (HashMap Text AttributeValue)

The primary key of the item where the operation stopped, inclusive of the previous result set. Use this value to start a new operation, excluding this value in the new request. If LastEvaluatedKey is empty, then the "last page" of results has been processed and there is no more data to be retrieved. If LastEvaluatedKey is not empty, it does not necessarily mean that there is more data in the result set. The only way to know when you have reached the end of the result set is when LastEvaluatedKey is empty.

srScannedCount :: Lens' ScanResponse (Maybe Int)

The number of items evaluated, before any ScanFilter is applied. A high ScannedCount value with few, or no, Count results indicates an inefficient Scan operation. For more information, see Count and ScannedCount in the Amazon DynamoDB Developer Guide. If you did not use a filter in the request, then ScannedCount is the same as Count.