Network.AWS.DynamoDB.UpdateTable

Description

Updates the provisioned throughput for the given table. Setting the throughput for a table helps you manage performance and is part of the provisioned throughput feature of DynamoDB.

The provisioned throughput values can be upgraded or downgraded based on the maximums and minimums listed in the Limits section in the Amazon DynamoDBDeveloper Guide.

The table must be in the ACTIVE state for this operation to succeed. UpdateTable is an asynchronous operation; while executing the operation, the table is in the UPDATING state. While the table is in the UPDATING state, the table still has the provisioned throughput from before the call. The new provisioned throughput setting is in effect only when the table returns to the ACTIVE state after the UpdateTable operation.

You cannot add, modify or delete indexes using UpdateTable. Indexes can only be defined at table creation time.

http://docs.aws.amazon.com/amazondynamodb/latest/APIReference/API_UpdateTable.html

Request

data UpdateTable

Request constructor

updateTable

Arguments

:: Text
utTableName

-> UpdateTable

UpdateTable constructor.

The fields accessible through corresponding lenses are:

utGlobalSecondaryIndexUpdates :: [GlobalSecondaryIndexUpdate]

utProvisionedThroughput :: Maybe ProvisionedThroughput

utTableName :: Text

Request lenses

utGlobalSecondaryIndexUpdates :: Lens' UpdateTable [GlobalSecondaryIndexUpdate]

An array of one or more global secondary indexes on the table, together with provisioned throughput settings for each index.

utProvisionedThroughput :: Lens' UpdateTable (Maybe ProvisionedThroughput)

utTableName :: Lens' UpdateTable Text

The name of the table to be updated.

Response

data UpdateTableResponse

Response constructor

updateTableResponse :: UpdateTableResponse

UpdateTableResponse constructor.

The fields accessible through corresponding lenses are:

utrTableDescription :: Maybe TableDescription

Response lenses

utrTableDescription :: Lens' UpdateTableResponse (Maybe TableDescription)