Network.AWS.DynamoDB.V2012_08_10.DeleteTable

Description

The DeleteTable operation deletes a table and all of its items. After a DeleteTable request, the specified table is in the DELETING state until DynamoDB completes the deletion. If the table is in the ACTIVE state, you can delete it. If a table is in CREATING or UPDATING states, then DynamoDB returns a ResourceInUseException. If the specified table does not exist, DynamoDB returns a ResourceNotFoundException. If table is already in the DELETING state, no error is returned. DynamoDB might continue to accept data read and write operations, such as GetItem and PutItem, on a table in the DELETING state until the table deletion is complete. When you delete a table, any indexes on that table are also deleted. Use the DescribeTable API to check the status of the table. Delete a Table This example deletes the Reply table. { TableDescription: { ItemCount: 0, ProvisionedThroughput: { NumberOfDecreasesToday: 0, ReadCapacityUnits: 5, WriteCapacityUnits: 5 }, TableName: Reply, TableSizeBytes: 0, TableStatus: DELETING } }.

Request

data DeleteTable

Represents the input of a DeleteTable operation.

Request constructor

mkDeleteTable

Arguments

:: Text
dtTableName

-> DeleteTable

Smart constructor for the minimum required parameters to construct a valid DeleteTable request.

The fields accessible through corresponding lenses are:

TableName :: Text

Request lenses

dtTableName :: Lens' DeleteTable Text

The name of the table to delete.

Response

data DeleteTableResponse

Represents the output of a DeleteTable operation.

Response constructor

mkDeleteTableResponse :: DeleteTableResponse

Smart constructor for the minimum required parameters to construct a valid DeleteTableResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

TableDescription :: Maybe TableDescription

Response lenses

dtrTableDescription :: Lens' DeleteTableResponse (Maybe TableDescription)

Represents the properties of a table.