Network.AWS.DynamoDB.V2012_08_10.DescribeTable

Description

Returns information about the table, including the current status of the table, when it was created, the primary key schema, and any indexes on the table. Describe a Table This example describes the Thread table. { Table: { AttributeDefinitions: [ { AttributeName: ForumName, AttributeType: S }, { AttributeName: LastPostDateTime, AttributeType: S }, { AttributeName: Subject, AttributeType: S } ], CreationDateTime: 1.363729002358E9, ItemCount: 0, KeySchema: [ { AttributeName: ForumName, KeyType: HASH }, { AttributeName: Subject, KeyType: RANGE } ], LocalSecondaryIndexes: [ { IndexName: LastPostIndex, IndexSizeBytes: 0, ItemCount: 0, KeySchema: [ { AttributeName: ForumName, KeyType: HASH }, { AttributeName: LastPostDateTime, KeyType: RANGE } ], Projection: { ProjectionType: KEYS_ONLY } } ], ProvisionedThroughput: { NumberOfDecreasesToday: 0, ReadCapacityUnits: 5, WriteCapacityUnits: 5 }, TableName: Thread, TableSizeBytes: 0, TableStatus: ACTIVE } }.

Documentation

data DescribeTable

Constructors

DescribeTable

Fields

_dtjTableName :: Text

The name of the table to describe.

dtjTableName :: Iso' DescribeTable Text

data DescribeTableResponse

Constructors

DescribeTableResponse

Fields

_dtpTable :: Maybe TableDescription

Represents the properties of a table.

dtpTable :: Iso' DescribeTableResponse (Maybe TableDescription)