Network.AWS.DynamoDB.CreateTable

Description

The CreateTable operation adds a new table to your account. In an AWS account, table names must be unique within each region. That is, you can have two tables with same name if you create the tables in different regions.

CreateTable is an asynchronous operation. Upon receiving a CreateTable request, DynamoDB immediately returns a response with a TableStatus of CREATING . After the table is created, DynamoDB sets the TableStatus to ACTIVE . You can perform read and write operations only on an ACTIVE table.

You can optionally define secondary indexes on the new table, as part of the CreateTable operation. If you want to create multiple tables with secondary indexes on them, you must create the tables sequentially. Only one table with secondary indexes can be in the CREATING state at any given time.

You can use the DescribeTable action to check the table status.

Creating a Request

createTable

Arguments

:: Text
ctTableName

-> NonEmpty KeySchemaElement
ctKeySchema

-> ProvisionedThroughput
ctProvisionedThroughput

-> CreateTable

Creates a value of CreateTable with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ctGlobalSecondaryIndexes - One or more global secondary indexes (the maximum is five) to be created on the table. Each global secondary index in the array includes the following: * IndexName - The name of the global secondary index. Must be unique only for this table. * KeySchema - Specifies the key schema for the global secondary index. * Projection - Specifies attributes that are copied (projected) from the table into the index. These are in addition to the primary key attributes and index key attributes, which are automatically projected. Each attribute specification is composed of: * ProjectionType - One of the following: * KEYS_ONLY - Only the index and primary keys are projected into the index. * INCLUDE - Only the specified table attributes are projected into the index. The list of projected attributes are in NonKeyAttributes . * ALL - All of the table attributes are projected into the index. * NonKeyAttributes - A list of one or more non-key attribute names that are projected into the secondary index. The total count of attributes provided in NonKeyAttributes , summed across all of the secondary indexes, must not exceed 20. If you project the same attribute into two different indexes, this counts as two distinct attributes when determining the total. * ProvisionedThroughput - The provisioned throughput settings for the global secondary index, consisting of read and write capacity units.

ctLocalSecondaryIndexes - One or more local secondary indexes (the maximum is five) to be created on the table. Each index is scoped to a given partition key value. There is a 10 GB size limit per partition key value; otherwise, the size of a local secondary index is unconstrained. Each local secondary index in the array includes the following: * IndexName - The name of the local secondary index. Must be unique only for this table. * KeySchema - Specifies the key schema for the local secondary index. The key schema must begin with the same partition key as the table. * Projection - Specifies attributes that are copied (projected) from the table into the index. These are in addition to the primary key attributes and index key attributes, which are automatically projected. Each attribute specification is composed of: * ProjectionType - One of the following: * KEYS_ONLY - Only the index and primary keys are projected into the index. * INCLUDE - Only the specified table attributes are projected into the index. The list of projected attributes are in NonKeyAttributes . * ALL - All of the table attributes are projected into the index. * NonKeyAttributes - A list of one or more non-key attribute names that are projected into the secondary index. The total count of attributes provided in NonKeyAttributes , summed across all of the secondary indexes, must not exceed 20. If you project the same attribute into two different indexes, this counts as two distinct attributes when determining the total.

ctStreamSpecification - The settings for DynamoDB Streams on the table. These settings consist of: * StreamEnabled - Indicates whether Streams is to be enabled (true) or disabled (false). * StreamViewType - When an item in the table is modified, StreamViewType determines what information is written to the table's stream. Valid values for StreamViewType are: * KEYS_ONLY - Only the key attributes of the modified item are written to the stream. * NEW_IMAGE - The entire item, as it appears after it was modified, is written to the stream. * OLD_IMAGE - The entire item, as it appeared before it was modified, is written to the stream. * NEW_AND_OLD_IMAGES - Both the new and the old item images of the item are written to the stream.

ctAttributeDefinitions - An array of attributes that describe the key schema for the table and indexes.

ctTableName - The name of the table to create.

ctKeySchema - Specifies the attributes that make up the primary key for a table or an index. The attributes in KeySchema must also be defined in the AttributeDefinitions array. For more information, see Data Model in the Amazon DynamoDB Developer Guide . Each KeySchemaElement in the array is composed of: * AttributeName - The name of this key attribute. * KeyType - The role that the key attribute will assume: * HASH - partition key * RANGE - sort key For a simple primary key (partition key), you must provide exactly one element with a KeyType of HASH . For a composite primary key (partition key and sort key), you must provide exactly two elements, in this order: The first element must have a KeyType of HASH , and the second element must have a KeyType of RANGE . For more information, see Specifying the Primary Key in the Amazon DynamoDB Developer Guide .

ctProvisionedThroughput - Represents the provisioned throughput settings for a specified table or index. The settings can be modified using the UpdateTable operation. For current minimum and maximum provisioned throughput values, see Limits in the Amazon DynamoDB Developer Guide .

data CreateTable

Represents the input of a CreateTable operation.

See: createTable smart constructor.

Request Lenses

ctGlobalSecondaryIndexes :: Lens' CreateTable [GlobalSecondaryIndex]

One or more global secondary indexes (the maximum is five) to be created on the table. Each global secondary index in the array includes the following: * IndexName - The name of the global secondary index. Must be unique only for this table. * KeySchema - Specifies the key schema for the global secondary index. * Projection - Specifies attributes that are copied (projected) from the table into the index. These are in addition to the primary key attributes and index key attributes, which are automatically projected. Each attribute specification is composed of: * ProjectionType - One of the following: * KEYS_ONLY - Only the index and primary keys are projected into the index. * INCLUDE - Only the specified table attributes are projected into the index. The list of projected attributes are in NonKeyAttributes . * ALL - All of the table attributes are projected into the index. * NonKeyAttributes - A list of one or more non-key attribute names that are projected into the secondary index. The total count of attributes provided in NonKeyAttributes , summed across all of the secondary indexes, must not exceed 20. If you project the same attribute into two different indexes, this counts as two distinct attributes when determining the total. * ProvisionedThroughput - The provisioned throughput settings for the global secondary index, consisting of read and write capacity units.

ctLocalSecondaryIndexes :: Lens' CreateTable [LocalSecondaryIndex]

One or more local secondary indexes (the maximum is five) to be created on the table. Each index is scoped to a given partition key value. There is a 10 GB size limit per partition key value; otherwise, the size of a local secondary index is unconstrained. Each local secondary index in the array includes the following: * IndexName - The name of the local secondary index. Must be unique only for this table. * KeySchema - Specifies the key schema for the local secondary index. The key schema must begin with the same partition key as the table. * Projection - Specifies attributes that are copied (projected) from the table into the index. These are in addition to the primary key attributes and index key attributes, which are automatically projected. Each attribute specification is composed of: * ProjectionType - One of the following: * KEYS_ONLY - Only the index and primary keys are projected into the index. * INCLUDE - Only the specified table attributes are projected into the index. The list of projected attributes are in NonKeyAttributes . * ALL - All of the table attributes are projected into the index. * NonKeyAttributes - A list of one or more non-key attribute names that are projected into the secondary index. The total count of attributes provided in NonKeyAttributes , summed across all of the secondary indexes, must not exceed 20. If you project the same attribute into two different indexes, this counts as two distinct attributes when determining the total.

ctStreamSpecification :: Lens' CreateTable (Maybe StreamSpecification)

The settings for DynamoDB Streams on the table. These settings consist of: * StreamEnabled - Indicates whether Streams is to be enabled (true) or disabled (false). * StreamViewType - When an item in the table is modified, StreamViewType determines what information is written to the table's stream. Valid values for StreamViewType are: * KEYS_ONLY - Only the key attributes of the modified item are written to the stream. * NEW_IMAGE - The entire item, as it appears after it was modified, is written to the stream. * OLD_IMAGE - The entire item, as it appeared before it was modified, is written to the stream. * NEW_AND_OLD_IMAGES - Both the new and the old item images of the item are written to the stream.

ctAttributeDefinitions :: Lens' CreateTable [AttributeDefinition]

An array of attributes that describe the key schema for the table and indexes.

ctTableName :: Lens' CreateTable Text

The name of the table to create.

ctKeySchema :: Lens' CreateTable (NonEmpty KeySchemaElement)

Specifies the attributes that make up the primary key for a table or an index. The attributes in KeySchema must also be defined in the AttributeDefinitions array. For more information, see Data Model in the Amazon DynamoDB Developer Guide . Each KeySchemaElement in the array is composed of: * AttributeName - The name of this key attribute. * KeyType - The role that the key attribute will assume: * HASH - partition key * RANGE - sort key For a simple primary key (partition key), you must provide exactly one element with a KeyType of HASH . For a composite primary key (partition key and sort key), you must provide exactly two elements, in this order: The first element must have a KeyType of HASH , and the second element must have a KeyType of RANGE . For more information, see Specifying the Primary Key in the Amazon DynamoDB Developer Guide .

ctProvisionedThroughput :: Lens' CreateTable ProvisionedThroughput

Represents the provisioned throughput settings for a specified table or index. The settings can be modified using the UpdateTable operation. For current minimum and maximum provisioned throughput values, see Limits in the Amazon DynamoDB Developer Guide .

Destructuring the Response

createTableResponse

Arguments

:: Int
ctrsResponseStatus

-> CreateTableResponse

Creates a value of CreateTableResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ctrsTableDescription - Represents the properties of the table.

ctrsResponseStatus - -- | The response status code.

data CreateTableResponse

Represents the output of a CreateTable operation.

See: createTableResponse smart constructor.

Response Lenses

ctrsTableDescription :: Lens' CreateTableResponse (Maybe TableDescription)

Represents the properties of the table.

ctrsResponseStatus :: Lens' CreateTableResponse Int

- | The response status code.