Network.AWS.DynamoDB.CreateTable

Description

The CreateTable operation adds a new table to your account. In an AWS account, table names must be unique within each region. That is, you can have two tables with same name if you create the tables in different regions. CreateTable is an asynchronous operation. Upon receiving a CreateTable request, DynamoDB immediately returns a response with a TableStatus of CREATING. After the table is created, DynamoDB sets the TableStatus to ACTIVE. You can perform read and write operations only on an ACTIVE table. If you want to create multiple tables with secondary indexes on them, you must create them sequentially. Only one table with secondary indexes can be in the CREATING state at any given time. You can use the DescribeTable API to check the table status.

http://docs.aws.amazon.com/amazondynamodb/latest/APIReference/API_CreateTable.html

Request

data CreateTable

Request constructor

createTable

Arguments

:: Text
ctTableName

-> NonEmpty KeySchemaElement
ctKeySchema

-> ProvisionedThroughput
ctProvisionedThroughput

-> CreateTable

CreateTable constructor.

The fields accessible through corresponding lenses are:

ctAttributeDefinitions :: [AttributeDefinition]

ctGlobalSecondaryIndexes :: [GlobalSecondaryIndex]

ctKeySchema :: NonEmpty KeySchemaElement

ctLocalSecondaryIndexes :: [LocalSecondaryIndex]

ctProvisionedThroughput :: ProvisionedThroughput

ctTableName :: Text

Request lenses

ctAttributeDefinitions :: Lens' CreateTable [AttributeDefinition]

An array of attributes that describe the key schema for the table and indexes.

ctGlobalSecondaryIndexes :: Lens' CreateTable [GlobalSecondaryIndex]

One or more global secondary indexes (the maximum is five) to be created on the table. Each global secondary index in the array includes the following: IndexName - The name of the global secondary index. Must be unique only for this table. KeySchema - Specifies the key schema for the global secondary index. Projection - Specifies attributes that are copied (projected) from the table into the index. These are in addition to the primary key attributes and index key attributes, which are automatically projected. Each attribute specification is composed of: ProjectionType - One of the following: KEYS_ONLY - Only the index and primary keys are projected into the index. INCLUDE - Only the specified table attributes are projected into the index. The list of projected attributes are in NonKeyAttributes. ALL - All of the table attributes are projected into the index. NonKeyAttributes - A list of one or more non-key attribute names that are projected into the secondary index. The total count of attributes specified in NonKeyAttributes, summed across all of the secondary indexes, must not exceed 20. If you project the same attribute into two different indexes, this counts as two distinct attributes when determining the total. ProvisionedThroughput - The provisioned throughput settings for the global secondary index, consisting of read and write capacity units.

ctKeySchema :: Lens' CreateTable (NonEmpty KeySchemaElement)

Specifies the attributes that make up the primary key for a table or an index. The attributes in KeySchema must also be defined in the AttributeDefinitions array. For more information, see Data Model in the Amazon DynamoDB Developer Guide. Each KeySchemaElement in the array is composed of: AttributeName - The name of this key attribute. KeyType - Determines whether the key attribute is HASH or RANGE. For a primary key that consists of a hash attribute, you must specify exactly one element with a KeyType of HASH. For a primary key that consists of hash and range attributes, you must specify exactly two elements, in this order: The first element must have a KeyType of HASH, and the second element must have a KeyType of RANGE. For more information, see Specifying the Primary Key in the Amazon DynamoDB Developer Guide.

ctLocalSecondaryIndexes :: Lens' CreateTable [LocalSecondaryIndex]

One or more local secondary indexes (the maximum is five) to be created on the table. Each index is scoped to a given hash key value. There is a 10 GB size limit per hash key; otherwise, the size of a local secondary index is unconstrained. Each local secondary index in the array includes the following: IndexName - The name of the local secondary index. Must be unique only for this table. KeySchema - Specifies the key schema for the local secondary index. The key schema must begin with the same hash key attribute as the table. Projection - Specifies attributes that are copied (projected) from the table into the index. These are in addition to the primary key attributes and index key attributes, which are automatically projected. Each attribute specification is composed of: ProjectionType - One of the following: KEYS_ONLY - Only the index and primary keys are projected into the index. INCLUDE - Only the specified table attributes are projected into the index. The list of projected attributes are in NonKeyAttributes. ALL - All of the table attributes are projected into the index. NonKeyAttributes - A list of one or more non-key attribute names that are projected into the secondary index. The total count of attributes specified in NonKeyAttributes, summed across all of the secondary indexes, must not exceed 20. If you project the same attribute into two different indexes, this counts as two distinct attributes when determining the total.

ctProvisionedThroughput :: Lens' CreateTable ProvisionedThroughput

ctTableName :: Lens' CreateTable Text

The name of the table to create.

Response

data CreateTableResponse

Response constructor

createTableResponse :: CreateTableResponse

CreateTableResponse constructor.

The fields accessible through corresponding lenses are:

ctrTableDescription :: Maybe TableDescription

Response lenses

ctrTableDescription :: Lens' CreateTableResponse (Maybe TableDescription)