Network.AWS.DynamoDB.DeleteItem

Description

Deletes a single item in a table by primary key. You can perform a conditional delete operation that deletes the item if it exists, or if it has an expected attribute value.

In addition to deleting an item, you can also return the item's attribute values in the same operation, using the ReturnValues parameter.

Unless you specify conditions, the DeleteItem is an idempotent operation; running it multiple times on the same item or attribute does not result in an error response.

Conditional deletes are useful for deleting items only if specific conditions are met. If those conditions are met, DynamoDB performs the delete. Otherwise, the item is not deleted.

Creating a Request

deleteItem

Arguments

:: Text
diTableName

-> DeleteItem

Creates a value of DeleteItem with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

diExpressionAttributeNames - One or more substitution tokens for attribute names in an expression. The following are some use cases for using ExpressionAttributeNames : * To access an attribute whose name conflicts with a DynamoDB reserved word. * To create a placeholder for repeating occurrences of an attribute name in an expression. * To prevent special characters in an attribute name from being misinterpreted in an expression. Use the # character in an expression to dereference an attribute name. For example, consider the following attribute name: * Percentile The name of this attribute conflicts with a reserved word, so it cannot be used directly in an expression. (For the complete list of reserved words, see Reserved Words in the Amazon DynamoDB Developer Guide ). To work around this, you could specify the following for ExpressionAttributeNames : * {"#P":Percentile} You could then use this substitution in an expression, as in this example: * #P = :val For more information on expression attribute names, see Accessing Item Attributes in the Amazon DynamoDB Developer Guide .

diReturnValues - Use ReturnValues if you want to get the item attributes as they appeared before they were deleted. For DeleteItem , the valid values are: * NONE - If ReturnValues is not specified, or if its value is NONE , then nothing is returned. (This setting is the default for ReturnValues .) * ALL_OLD - The content of the old item is returned.

diExpressionAttributeValues - One or more values that can be substituted in an expression. Use the : (colon) character in an expression to dereference an attribute value. For example, suppose that you wanted to check whether the value of the ProductStatus attribute was one of the following: Available | Backordered | Discontinued You would first need to specify ExpressionAttributeValues as follows: { ":avail":{S:Available}, ":back":{S:Backordered}, ":disc":{S:Discontinued} } You could then use these values in an expression, such as this: ProductStatus IN (:avail, :back, :disc) For more information on expression attribute values, see Specifying Conditions in the Amazon DynamoDB Developer Guide .

diReturnConsumedCapacity - Undocumented member.

diReturnItemCollectionMetrics - Determines whether item collection metrics are returned. If set to SIZE , the response includes statistics about item collections, if any, that were modified during the operation are returned in the response. If set to NONE (the default), no statistics are returned.

diConditionExpression - A condition that must be satisfied in order for a conditional DeleteItem to succeed. An expression can contain any of the following: * Functions: attribute_exists | attribute_not_exists | attribute_type | contains | begins_with | size These function names are case-sensitive. * Comparison operators: = | <> | | | |= | BETWEEN | IN * Logical operators: AND | OR | NOT For more information on condition expressions, see Specifying Conditions in the Amazon DynamoDB Developer Guide .

diConditionalOperator - This is a legacy parameter. Use ConditionExpression instead. For more information, see ConditionalOperator in the Amazon DynamoDB Developer Guide .

diExpected - This is a legacy parameter. Use ConditionExpression instead. For more information, see Expected in the Amazon DynamoDB Developer Guide .

diTableName - The name of the table from which to delete the item.

diKey - A map of attribute names to AttributeValue objects, representing the primary key of the item to delete. For the primary key, you must provide all of the attributes. For example, with a simple primary key, you only need to provide a value for the partition key. For a composite primary key, you must provide values for both the partition key and the sort key.

data DeleteItem

Represents the input of a DeleteItem operation.

See: deleteItem smart constructor.

Request Lenses

diExpressionAttributeNames :: Lens' DeleteItem (HashMap Text Text)

One or more substitution tokens for attribute names in an expression. The following are some use cases for using ExpressionAttributeNames : * To access an attribute whose name conflicts with a DynamoDB reserved word. * To create a placeholder for repeating occurrences of an attribute name in an expression. * To prevent special characters in an attribute name from being misinterpreted in an expression. Use the # character in an expression to dereference an attribute name. For example, consider the following attribute name: * Percentile The name of this attribute conflicts with a reserved word, so it cannot be used directly in an expression. (For the complete list of reserved words, see Reserved Words in the Amazon DynamoDB Developer Guide ). To work around this, you could specify the following for ExpressionAttributeNames : * {"#P":Percentile} You could then use this substitution in an expression, as in this example: * #P = :val For more information on expression attribute names, see Accessing Item Attributes in the Amazon DynamoDB Developer Guide .

diReturnValues :: Lens' DeleteItem (Maybe ReturnValue)

Use ReturnValues if you want to get the item attributes as they appeared before they were deleted. For DeleteItem , the valid values are: * NONE - If ReturnValues is not specified, or if its value is NONE , then nothing is returned. (This setting is the default for ReturnValues .) * ALL_OLD - The content of the old item is returned.

diExpressionAttributeValues :: Lens' DeleteItem (HashMap Text AttributeValue)

One or more values that can be substituted in an expression. Use the : (colon) character in an expression to dereference an attribute value. For example, suppose that you wanted to check whether the value of the ProductStatus attribute was one of the following: Available | Backordered | Discontinued You would first need to specify ExpressionAttributeValues as follows: { ":avail":{S:Available}, ":back":{S:Backordered}, ":disc":{S:Discontinued} } You could then use these values in an expression, such as this: ProductStatus IN (:avail, :back, :disc) For more information on expression attribute values, see Specifying Conditions in the Amazon DynamoDB Developer Guide .

diReturnConsumedCapacity :: Lens' DeleteItem (Maybe ReturnConsumedCapacity)

Undocumented member.

diReturnItemCollectionMetrics :: Lens' DeleteItem (Maybe ReturnItemCollectionMetrics)

Determines whether item collection metrics are returned. If set to SIZE , the response includes statistics about item collections, if any, that were modified during the operation are returned in the response. If set to NONE (the default), no statistics are returned.

diConditionExpression :: Lens' DeleteItem (Maybe Text)

A condition that must be satisfied in order for a conditional DeleteItem to succeed. An expression can contain any of the following: * Functions: attribute_exists | attribute_not_exists | attribute_type | contains | begins_with | size These function names are case-sensitive. * Comparison operators: = | <> | | | |= | BETWEEN | IN * Logical operators: AND | OR | NOT For more information on condition expressions, see Specifying Conditions in the Amazon DynamoDB Developer Guide .

diConditionalOperator :: Lens' DeleteItem (Maybe ConditionalOperator)

This is a legacy parameter. Use ConditionExpression instead. For more information, see ConditionalOperator in the Amazon DynamoDB Developer Guide .

diExpected :: Lens' DeleteItem (HashMap Text ExpectedAttributeValue)

This is a legacy parameter. Use ConditionExpression instead. For more information, see Expected in the Amazon DynamoDB Developer Guide .

diTableName :: Lens' DeleteItem Text

The name of the table from which to delete the item.

diKey :: Lens' DeleteItem (HashMap Text AttributeValue)

A map of attribute names to AttributeValue objects, representing the primary key of the item to delete. For the primary key, you must provide all of the attributes. For example, with a simple primary key, you only need to provide a value for the partition key. For a composite primary key, you must provide values for both the partition key and the sort key.

Destructuring the Response

deleteItemResponse

Arguments

:: Int
dirsResponseStatus

-> DeleteItemResponse

Creates a value of DeleteItemResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dirsItemCollectionMetrics - Information about item collections, if any, that were affected by the DeleteItem operation. ItemCollectionMetrics is only returned if the ReturnItemCollectionMetrics parameter was specified. If the table does not have any local secondary indexes, this information is not returned in the response. Each ItemCollectionMetrics element consists of: * ItemCollectionKey - The partition key value of the item collection. This is the same as the partition key value of the item itself. * SizeEstimateRange - An estimate of item collection size, in gigabytes. This value is a two-element array containing a lower bound and an upper bound for the estimate. The estimate includes the size of all the items in the table, plus the size of all attributes projected into all of the local secondary indexes on that table. Use this estimate to measure whether a local secondary index is approaching its size limit. The estimate is subject to change over time; therefore, do not rely on the precision or accuracy of the estimate.

dirsConsumedCapacity - The capacity units consumed by the DeleteItem operation. The data returned includes the total provisioned throughput consumed, along with statistics for the table and any indexes involved in the operation. ConsumedCapacity is only returned if the ReturnConsumedCapacity parameter was specified. For more information, see Provisioned Throughput in the Amazon DynamoDB Developer Guide .

dirsAttributes - A map of attribute names to AttributeValue objects, representing the item as it appeared before the DeleteItem operation. This map appears in the response only if ReturnValues was specified as ALL_OLD in the request.

dirsResponseStatus - -- | The response status code.

data DeleteItemResponse

Represents the output of a DeleteItem operation.

See: deleteItemResponse smart constructor.

Response Lenses

dirsItemCollectionMetrics :: Lens' DeleteItemResponse (Maybe ItemCollectionMetrics)

Information about item collections, if any, that were affected by the DeleteItem operation. ItemCollectionMetrics is only returned if the ReturnItemCollectionMetrics parameter was specified. If the table does not have any local secondary indexes, this information is not returned in the response. Each ItemCollectionMetrics element consists of: * ItemCollectionKey - The partition key value of the item collection. This is the same as the partition key value of the item itself. * SizeEstimateRange - An estimate of item collection size, in gigabytes. This value is a two-element array containing a lower bound and an upper bound for the estimate. The estimate includes the size of all the items in the table, plus the size of all attributes projected into all of the local secondary indexes on that table. Use this estimate to measure whether a local secondary index is approaching its size limit. The estimate is subject to change over time; therefore, do not rely on the precision or accuracy of the estimate.

dirsConsumedCapacity :: Lens' DeleteItemResponse (Maybe ConsumedCapacity)

The capacity units consumed by the DeleteItem operation. The data returned includes the total provisioned throughput consumed, along with statistics for the table and any indexes involved in the operation. ConsumedCapacity is only returned if the ReturnConsumedCapacity parameter was specified. For more information, see Provisioned Throughput in the Amazon DynamoDB Developer Guide .

dirsAttributes :: Lens' DeleteItemResponse (HashMap Text AttributeValue)

A map of attribute names to AttributeValue objects, representing the item as it appeared before the DeleteItem operation. This map appears in the response only if ReturnValues was specified as ALL_OLD in the request.

dirsResponseStatus :: Lens' DeleteItemResponse Int

- | The response status code.