Network.AWS.DynamoDB.DescribeTable

Description

Returns information about the table, including the current status of the table, when it was created, the primary key schema, and any indexes on the table.

Creating a Request

describeTable

Arguments

:: Text
dTableName

-> DescribeTable

Creates a value of DescribeTable with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dTableName - The name of the table to describe.

data DescribeTable

Represents the input of a DescribeTable operation.

See: describeTable smart constructor.

Request Lenses

dTableName :: Lens' DescribeTable Text

The name of the table to describe.

Destructuring the Response

describeTableResponse

Arguments

:: Int
drsResponseStatus

-> DescribeTableResponse

Creates a value of DescribeTableResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drsTable - The properties of the table.

drsResponseStatus - -- | The response status code.

data DescribeTableResponse

Represents the output of a DescribeTable operation.

See: describeTableResponse smart constructor.

Response Lenses

drsTable :: Lens' DescribeTableResponse (Maybe TableDescription)

The properties of the table.

drsResponseStatus :: Lens' DescribeTableResponse Int

- | The response status code.