Network.AWS.DynamoDB.DescribeTable

Description

Returns information about the table, including the current status of the table, when it was created, the primary key schema, and any indexes on the table.

If you issue a DescribeTable request immediately after a CreateTable request, DynamoDB might return a ResourceNotFoundException. This is because DescribeTable uses an eventually consistent query, and the metadata for your table might not be available at that moment. Wait for a few seconds, and then try the DescribeTable request again.

http://docs.aws.amazon.com/amazondynamodb/latest/APIReference/API_DescribeTable.html

Request

data DescribeTable

Request constructor

describeTable

Arguments

:: Text
dt1TableName

-> DescribeTable

DescribeTable constructor.

The fields accessible through corresponding lenses are:

dt1TableName :: Text

Request lenses

dt1TableName :: Lens' DescribeTable Text

The name of the table to describe.

Response

data DescribeTableResponse

Response constructor

describeTableResponse

Arguments

:: TableDescription
dtrTable

-> DescribeTableResponse

DescribeTableResponse constructor.

The fields accessible through corresponding lenses are:

dtrTable :: TableDescription

Response lenses

dtrTable :: Lens' DescribeTableResponse TableDescription