Network.AWS.DynamoDB.GetItem

Description

The GetItem operation returns a set of attributes for the item with the given primary key. If there is no matching item, GetItem does not return any data and there will be no Item element in the response.

GetItem provides an eventually consistent read by default. If your application requires a strongly consistent read, set ConsistentRead to true . Although a strongly consistent read might take more time than an eventually consistent read, it always returns the last updated value.

Creating a Request

getItem

Arguments

:: Text
giTableName

-> GetItem

Creates a value of GetItem with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

giProjectionExpression - A string that identifies one or more attributes to retrieve from the table. These attributes can include scalars, sets, or elements of a JSON document. The attributes in the expression must be separated by commas. If no attribute names are specified, then all attributes will be returned. If any of the requested attributes are not found, they will not appear in the result. For more information, see Accessing Item Attributes in the Amazon DynamoDB Developer Guide .

giAttributesToGet - This is a legacy parameter. Use ProjectionExpression instead. For more information, see AttributesToGet in the Amazon DynamoDB Developer Guide .

giExpressionAttributeNames - One or more substitution tokens for attribute names in an expression. The following are some use cases for using ExpressionAttributeNames : * To access an attribute whose name conflicts with a DynamoDB reserved word. * To create a placeholder for repeating occurrences of an attribute name in an expression. * To prevent special characters in an attribute name from being misinterpreted in an expression. Use the # character in an expression to dereference an attribute name. For example, consider the following attribute name: * Percentile The name of this attribute conflicts with a reserved word, so it cannot be used directly in an expression. (For the complete list of reserved words, see Reserved Words in the Amazon DynamoDB Developer Guide ). To work around this, you could specify the following for ExpressionAttributeNames : * {"#P":Percentile} You could then use this substitution in an expression, as in this example: * #P = :val For more information on expression attribute names, see Accessing Item Attributes in the Amazon DynamoDB Developer Guide .

giConsistentRead - Determines the read consistency model: If set to true , then the operation uses strongly consistent reads; otherwise, the operation uses eventually consistent reads.

giReturnConsumedCapacity - Undocumented member.

giTableName - The name of the table containing the requested item.

giKey - A map of attribute names to AttributeValue objects, representing the primary key of the item to retrieve. For the primary key, you must provide all of the attributes. For example, with a simple primary key, you only need to provide a value for the partition key. For a composite primary key, you must provide values for both the partition key and the sort key.

data GetItem

Represents the input of a GetItem operation.

See: getItem smart constructor.

Request Lenses

giProjectionExpression :: Lens' GetItem (Maybe Text)

A string that identifies one or more attributes to retrieve from the table. These attributes can include scalars, sets, or elements of a JSON document. The attributes in the expression must be separated by commas. If no attribute names are specified, then all attributes will be returned. If any of the requested attributes are not found, they will not appear in the result. For more information, see Accessing Item Attributes in the Amazon DynamoDB Developer Guide .

giAttributesToGet :: Lens' GetItem (Maybe (NonEmpty Text))

This is a legacy parameter. Use ProjectionExpression instead. For more information, see AttributesToGet in the Amazon DynamoDB Developer Guide .

giExpressionAttributeNames :: Lens' GetItem (HashMap Text Text)

One or more substitution tokens for attribute names in an expression. The following are some use cases for using ExpressionAttributeNames : * To access an attribute whose name conflicts with a DynamoDB reserved word. * To create a placeholder for repeating occurrences of an attribute name in an expression. * To prevent special characters in an attribute name from being misinterpreted in an expression. Use the # character in an expression to dereference an attribute name. For example, consider the following attribute name: * Percentile The name of this attribute conflicts with a reserved word, so it cannot be used directly in an expression. (For the complete list of reserved words, see Reserved Words in the Amazon DynamoDB Developer Guide ). To work around this, you could specify the following for ExpressionAttributeNames : * {"#P":Percentile} You could then use this substitution in an expression, as in this example: * #P = :val For more information on expression attribute names, see Accessing Item Attributes in the Amazon DynamoDB Developer Guide .

giConsistentRead :: Lens' GetItem (Maybe Bool)

Determines the read consistency model: If set to true , then the operation uses strongly consistent reads; otherwise, the operation uses eventually consistent reads.

giReturnConsumedCapacity :: Lens' GetItem (Maybe ReturnConsumedCapacity)

Undocumented member.

giTableName :: Lens' GetItem Text

The name of the table containing the requested item.

giKey :: Lens' GetItem (HashMap Text AttributeValue)

A map of attribute names to AttributeValue objects, representing the primary key of the item to retrieve. For the primary key, you must provide all of the attributes. For example, with a simple primary key, you only need to provide a value for the partition key. For a composite primary key, you must provide values for both the partition key and the sort key.

Destructuring the Response

getItemResponse

Arguments

:: Int
girsResponseStatus

-> GetItemResponse

Creates a value of GetItemResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

girsConsumedCapacity - The capacity units consumed by the GetItem operation. The data returned includes the total provisioned throughput consumed, along with statistics for the table and any indexes involved in the operation. ConsumedCapacity is only returned if the ReturnConsumedCapacity parameter was specified. For more information, see Provisioned Throughput in the Amazon DynamoDB Developer Guide .

girsItem - A map of attribute names to AttributeValue objects, as specified by ProjectionExpression .

girsResponseStatus - -- | The response status code.

data GetItemResponse

Represents the output of a GetItem operation.

See: getItemResponse smart constructor.

Response Lenses

girsConsumedCapacity :: Lens' GetItemResponse (Maybe ConsumedCapacity)

The capacity units consumed by the GetItem operation. The data returned includes the total provisioned throughput consumed, along with statistics for the table and any indexes involved in the operation. ConsumedCapacity is only returned if the ReturnConsumedCapacity parameter was specified. For more information, see Provisioned Throughput in the Amazon DynamoDB Developer Guide .

girsItem :: Lens' GetItemResponse (HashMap Text AttributeValue)

A map of attribute names to AttributeValue objects, as specified by ProjectionExpression .

girsResponseStatus :: Lens' GetItemResponse Int

- | The response status code.