Network.AWS.DynamoDB.GetItem

Description

The GetItem operation returns a set of attributes for the item with the given primary key. If there is no matching item, GetItem does not return any data.

GetItem provides an eventually consistent read by default. If your application requires a strongly consistent read, set ConsistentRead to true. Although a strongly consistent read might take more time than an eventually consistent read, it always returns the last updated value.

http://docs.aws.amazon.com/amazondynamodb/latest/APIReference/API_GetItem.html

Request

data GetItem

Request constructor

getItem

Arguments

:: Text
giTableName

-> NonEmpty Text
giAttributesToGet

-> GetItem

GetItem constructor.

The fields accessible through corresponding lenses are:

giAttributesToGet :: NonEmpty Text

giConsistentRead :: Maybe Bool

giExpressionAttributeNames :: HashMap Text Text

giKey :: HashMap Text AttributeValue

giProjectionExpression :: Maybe Text

giReturnConsumedCapacity :: Maybe ReturnConsumedCapacity

giTableName :: Text

Request lenses

giAttributesToGet :: Lens' GetItem (NonEmpty Text)

There is a newer parameter available. Use ProjectionExpression instead. Note that if you use AttributesToGet and ProjectionExpression at the same time, DynamoDB will return a ValidationException exception.

This parameter allows you to retrieve lists or maps; however, it cannot retrieve individual list or map elements.

The names of one or more attributes to retrieve. If no attribute names are specified, then all attributes will be returned. If any of the requested attributes are not found, they will not appear in the result.

Note that AttributesToGet has no effect on provisioned throughput consumption. DynamoDB determines capacity units consumed based on item size, not on the amount of data that is returned to an application.

giConsistentRead :: Lens' GetItem (Maybe Bool)

A value that if set to true, then the operation uses strongly consistent reads; otherwise, eventually consistent reads are used.

giExpressionAttributeNames :: Lens' GetItem (HashMap Text Text)

One or more substitution tokens for simplifying complex expressions. The following are some use cases for an ExpressionAttributeNames value:

To shorten an attribute name that is very long or unwieldy in an expression.

To create a placeholder for repeating occurrences of an attribute name in an expression.

To prevent special characters in an attribute name from being misinterpreted in an expression.

Use the # character in an expression to dereference an attribute name. For example, consider the following expression:

'order.customerInfo.LastName = Smith OR order.customerInfo.LastName =Jones'

Now suppose that you specified the following for ExpressionAttributeNames:

'{"n":"order.customerInfo.LastName"}'

The expression can now be simplified as follows:

'n = Jones'

giKey :: Lens' GetItem (HashMap Text AttributeValue)

A map of attribute names to AttributeValue objects, representing the primary key of the item to retrieve.

For the primary key, you must provide all of the attributes. For example, with a hash type primary key, you only need to specify the hash attribute. For a hash-and-range type primary key, you must specify both the hash attribute and the range attribute.

giProjectionExpression :: Lens' GetItem (Maybe Text)

One or more attributes to retrieve from the table. These attributes can include scalars, sets, or elements of a JSON document. The attributes in the expression must be separated by commas.

If no attribute names are specified, then all attributes will be returned. If any of the requested attributes are not found, they will not appear in the result.

giReturnConsumedCapacity :: Lens' GetItem (Maybe ReturnConsumedCapacity)

giTableName :: Lens' GetItem Text

The name of the table containing the requested item.

Response

data GetItemResponse

Response constructor

getItemResponse :: GetItemResponse

GetItemResponse constructor.

The fields accessible through corresponding lenses are:

girConsumedCapacity :: Maybe ConsumedCapacity

girItem :: HashMap Text AttributeValue

Response lenses

girConsumedCapacity :: Lens' GetItemResponse (Maybe ConsumedCapacity)

girItem :: Lens' GetItemResponse (HashMap Text AttributeValue)

A map of attribute names to AttributeValue objects, as specified by AttributesToGet.