Network.AWS.DynamoDB.GetItem

Description

The GetItem operation returns a set of attributes for the item with the given primary key. If there is no matching item, GetItem does not return any data.

GetItem provides an eventually consistent read by default. If your application requires a strongly consistent read, set ConsistentRead to true. Although a strongly consistent read might take more time than an eventually consistent read, it always returns the last updated value.

http://docs.aws.amazon.com/amazondynamodb/latest/APIReference/API_GetItem.html

Request

data GetItem

Request constructor

getItem

Arguments

:: Text
giTableName

-> NonEmpty Text
giAttributesToGet

-> GetItem

GetItem constructor.

The fields accessible through corresponding lenses are:

giAttributesToGet :: NonEmpty Text

giConsistentRead :: Maybe Bool

giExpressionAttributeNames :: HashMap Text Text

giKey :: HashMap Text AttributeValue

giProjectionExpression :: Maybe Text

giReturnConsumedCapacity :: Maybe ReturnConsumedCapacity

giTableName :: Text

Request lenses

giAttributesToGet :: Lens' GetItem (NonEmpty Text)

There is a newer parameter available. Use ProjectionExpression instead. Note that if you use AttributesToGet and ProjectionExpression at the same time, DynamoDB will return a ValidationException exception.

This parameter allows you to retrieve attributes of type List or Map; however, it cannot retrieve individual elements within a List or a Map.

The names of one or more attributes to retrieve. If no attribute names are provided, then all attributes will be returned. If any of the requested attributes are not found, they will not appear in the result.

Note that AttributesToGet has no effect on provisioned throughput consumption. DynamoDB determines capacity units consumed based on item size, not on the amount of data that is returned to an application.

giConsistentRead :: Lens' GetItem (Maybe Bool)

A value that if set to true, then the operation uses strongly consistent reads; otherwise, eventually consistent reads are used.

giExpressionAttributeNames :: Lens' GetItem (HashMap Text Text)

One or more substitution tokens for attribute names in an expression. The following are some use cases for using ExpressionAttributeNames:

To access an attribute whose name conflicts with a DynamoDB reserved word.

To create a placeholder for repeating occurrences of an attribute name in an expression.

To prevent special characters in an attribute name from being misinterpreted in an expression.

Use the # character in an expression to dereference an attribute name. For example, consider the following attribute name:

Percentile

The name of this attribute conflicts with a reserved word, so it cannot be used directly in an expression. (For the complete list of reserved words, go to Reserved Words in the Amazon DynamoDB Developer Guide). To work around this, you could specify the following for ExpressionAttributeNames:

'{"#P":Percentile}'

You could then use this substitution in an expression, as in this example:

'#P = :val'

Tokens that begin with the : character are expression attribute values, which are placeholders for the actual value at runtime.

For more information on expression attribute names, go to Accessing ItemAttributes in the Amazon DynamoDB Developer Guide.

giKey :: Lens' GetItem (HashMap Text AttributeValue)

A map of attribute names to AttributeValue objects, representing the primary key of the item to retrieve.

For the primary key, you must provide all of the attributes. For example, with a hash type primary key, you only need to provide the hash attribute. For a hash-and-range type primary key, you must provide both the hash attribute and the range attribute.

giProjectionExpression :: Lens' GetItem (Maybe Text)

A string that identifies one or more attributes to retrieve from the table. These attributes can include scalars, sets, or elements of a JSON document. The attributes in the expression must be separated by commas.

If no attribute names are specified, then all attributes will be returned. If any of the requested attributes are not found, they will not appear in the result.

For more information, go to Accessing Item Attributes in the Amazon DynamoDBDeveloper Guide.

giReturnConsumedCapacity :: Lens' GetItem (Maybe ReturnConsumedCapacity)

giTableName :: Lens' GetItem Text

The name of the table containing the requested item.

Response

data GetItemResponse

Response constructor

getItemResponse :: GetItemResponse

GetItemResponse constructor.

The fields accessible through corresponding lenses are:

girConsumedCapacity :: Maybe ConsumedCapacity

girItem :: HashMap Text AttributeValue

Response lenses

girConsumedCapacity :: Lens' GetItemResponse (Maybe ConsumedCapacity)

girItem :: Lens' GetItemResponse (HashMap Text AttributeValue)

A map of attribute names to AttributeValue objects, as specified by AttributesToGet.