Network.AWS.DynamoDB.PutItem

Description

Creates a new item, or replaces an old item with a new item. If an item that has the same primary key as the new item already exists in the specified table, the new item completely replaces the existing item. You can perform a conditional put operation (add a new item if one with the specified primary key doesn't exist), or replace an existing item if it has certain attribute values. You can return the item's attribute values in the same operation, using the ReturnValues parameter.

Important: This topic provides general information about the PutItem API.

For information on how to call the PutItem API using the AWS SDK in specific languages, see the following:

PutItem in the AWS Command Line Interface

PutItem in the AWS SDK for .NET

PutItem in the AWS SDK for C++

PutItem in the AWS SDK for Go

PutItem in the AWS SDK for Java

PutItem in the AWS SDK for JavaScript

PutItem in the AWS SDK for PHP V3

PutItem in the AWS SDK for Python

PutItem in the AWS SDK for Ruby V2

When you add an item, the primary key attribute(s) are the only required attributes. Attribute values cannot be null. String and Binary type attributes must have lengths greater than zero. Set type attributes cannot be empty. Requests with empty values will be rejected with a ValidationException exception.

For more information about PutItem , see Working with Items in the Amazon DynamoDB Developer Guide .

Creating a Request

putItem

Arguments

:: Text
piTableName

-> PutItem

Creates a value of PutItem with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

piExpressionAttributeNames - One or more substitution tokens for attribute names in an expression. The following are some use cases for using ExpressionAttributeNames : * To access an attribute whose name conflicts with a DynamoDB reserved word. * To create a placeholder for repeating occurrences of an attribute name in an expression. * To prevent special characters in an attribute name from being misinterpreted in an expression. Use the # character in an expression to dereference an attribute name. For example, consider the following attribute name: * Percentile The name of this attribute conflicts with a reserved word, so it cannot be used directly in an expression. (For the complete list of reserved words, see Reserved Words in the Amazon DynamoDB Developer Guide ). To work around this, you could specify the following for ExpressionAttributeNames : * {"#P":Percentile} You could then use this substitution in an expression, as in this example: * #P = :val For more information on expression attribute names, see Accessing Item Attributes in the Amazon DynamoDB Developer Guide .

piReturnValues - Use ReturnValues if you want to get the item attributes as they appeared before they were updated with the PutItem request. For PutItem , the valid values are: * NONE - If ReturnValues is not specified, or if its value is NONE , then nothing is returned. (This setting is the default for ReturnValues .) * ALL_OLD - If PutItem overwrote an attribute name-value pair, then the content of the old item is returned.

piExpressionAttributeValues - One or more values that can be substituted in an expression. Use the : (colon) character in an expression to dereference an attribute value. For example, suppose that you wanted to check whether the value of the ProductStatus attribute was one of the following: Available | Backordered | Discontinued You would first need to specify ExpressionAttributeValues as follows: { ":avail":{S:Available}, ":back":{S:Backordered}, ":disc":{S:Discontinued} } You could then use these values in an expression, such as this: ProductStatus IN (:avail, :back, :disc) For more information on expression attribute values, see Specifying Conditions in the Amazon DynamoDB Developer Guide .

piReturnConsumedCapacity - Undocumented member.

piReturnItemCollectionMetrics - Determines whether item collection metrics are returned. If set to SIZE , the response includes statistics about item collections, if any, that were modified during the operation are returned in the response. If set to NONE (the default), no statistics are returned.

piConditionExpression - A condition that must be satisfied in order for a conditional PutItem operation to succeed. An expression can contain any of the following: * Functions: attribute_exists | attribute_not_exists | attribute_type | contains | begins_with | size These function names are case-sensitive. * Comparison operators: = | <> | | | |= | BETWEEN | IN * Logical operators: AND | OR | NOT For more information on condition expressions, see Specifying Conditions in the Amazon DynamoDB Developer Guide .

piConditionalOperator - This is a legacy parameter. Use ConditionExpression instead. For more information, see ConditionalOperator in the Amazon DynamoDB Developer Guide .

piExpected - This is a legacy parameter. Use ConditionExpression instead. For more information, see Expected in the Amazon DynamoDB Developer Guide .

piTableName - The name of the table to contain the item.

piItem - A map of attribute namevalue pairs, one for each attribute. Only the primary key attributes are required; you can optionally provide other attribute name-value pairs for the item. You must provide all of the attributes for the primary key. For example, with a simple primary key, you only need to provide a value for the partition key. For a composite primary key, you must provide both values for both the partition key and the sort key. If you specify any attributes that are part of an index key, then the data types for those attributes must match those of the schema in the table's attribute definition. For more information about primary keys, see <http:docs.aws.amazon.comamazondynamodblatestdeveloperguideDataModel.html#DataModelPrimaryKey Primary Key> in the Amazon DynamoDB Developer Guide/ . Each element in the Item map is an AttributeValue object.

data PutItem

Represents the input of a PutItem operation.

See: putItem smart constructor.

Request Lenses

piExpressionAttributeNames :: Lens' PutItem (HashMap Text Text)

One or more substitution tokens for attribute names in an expression. The following are some use cases for using ExpressionAttributeNames : * To access an attribute whose name conflicts with a DynamoDB reserved word. * To create a placeholder for repeating occurrences of an attribute name in an expression. * To prevent special characters in an attribute name from being misinterpreted in an expression. Use the # character in an expression to dereference an attribute name. For example, consider the following attribute name: * Percentile The name of this attribute conflicts with a reserved word, so it cannot be used directly in an expression. (For the complete list of reserved words, see Reserved Words in the Amazon DynamoDB Developer Guide ). To work around this, you could specify the following for ExpressionAttributeNames : * {"#P":Percentile} You could then use this substitution in an expression, as in this example: * #P = :val For more information on expression attribute names, see Accessing Item Attributes in the Amazon DynamoDB Developer Guide .

piReturnValues :: Lens' PutItem (Maybe ReturnValue)

Use ReturnValues if you want to get the item attributes as they appeared before they were updated with the PutItem request. For PutItem , the valid values are: * NONE - If ReturnValues is not specified, or if its value is NONE , then nothing is returned. (This setting is the default for ReturnValues .) * ALL_OLD - If PutItem overwrote an attribute name-value pair, then the content of the old item is returned.

piExpressionAttributeValues :: Lens' PutItem (HashMap Text AttributeValue)

One or more values that can be substituted in an expression. Use the : (colon) character in an expression to dereference an attribute value. For example, suppose that you wanted to check whether the value of the ProductStatus attribute was one of the following: Available | Backordered | Discontinued You would first need to specify ExpressionAttributeValues as follows: { ":avail":{S:Available}, ":back":{S:Backordered}, ":disc":{S:Discontinued} } You could then use these values in an expression, such as this: ProductStatus IN (:avail, :back, :disc) For more information on expression attribute values, see Specifying Conditions in the Amazon DynamoDB Developer Guide .

piReturnConsumedCapacity :: Lens' PutItem (Maybe ReturnConsumedCapacity)

Undocumented member.

piReturnItemCollectionMetrics :: Lens' PutItem (Maybe ReturnItemCollectionMetrics)

Determines whether item collection metrics are returned. If set to SIZE , the response includes statistics about item collections, if any, that were modified during the operation are returned in the response. If set to NONE (the default), no statistics are returned.

piConditionExpression :: Lens' PutItem (Maybe Text)

A condition that must be satisfied in order for a conditional PutItem operation to succeed. An expression can contain any of the following: * Functions: attribute_exists | attribute_not_exists | attribute_type | contains | begins_with | size These function names are case-sensitive. * Comparison operators: = | <> | | | |= | BETWEEN | IN * Logical operators: AND | OR | NOT For more information on condition expressions, see Specifying Conditions in the Amazon DynamoDB Developer Guide .

piConditionalOperator :: Lens' PutItem (Maybe ConditionalOperator)

This is a legacy parameter. Use ConditionExpression instead. For more information, see ConditionalOperator in the Amazon DynamoDB Developer Guide .

piExpected :: Lens' PutItem (HashMap Text ExpectedAttributeValue)

This is a legacy parameter. Use ConditionExpression instead. For more information, see Expected in the Amazon DynamoDB Developer Guide .

piTableName :: Lens' PutItem Text

The name of the table to contain the item.

piItem :: Lens' PutItem (HashMap Text AttributeValue)

A map of attribute namevalue pairs, one for each attribute. Only the primary key attributes are required; you can optionally provide other attribute name-value pairs for the item. You must provide all of the attributes for the primary key. For example, with a simple primary key, you only need to provide a value for the partition key. For a composite primary key, you must provide both values for both the partition key and the sort key. If you specify any attributes that are part of an index key, then the data types for those attributes must match those of the schema in the table's attribute definition. For more information about primary keys, see <http:docs.aws.amazon.comamazondynamodblatestdeveloperguideDataModel.html#DataModelPrimaryKey Primary Key> in the Amazon DynamoDB Developer Guide/ . Each element in the Item map is an AttributeValue object.

Destructuring the Response

putItemResponse

Arguments

:: Int
pirsResponseStatus

-> PutItemResponse

Creates a value of PutItemResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pirsItemCollectionMetrics - Information about item collections, if any, that were affected by the PutItem operation. ItemCollectionMetrics is only returned if the ReturnItemCollectionMetrics parameter was specified. If the table does not have any local secondary indexes, this information is not returned in the response. Each ItemCollectionMetrics element consists of: * ItemCollectionKey - The partition key value of the item collection. This is the same as the partition key value of the item itself. * SizeEstimateRange - An estimate of item collection size, in gigabytes. This value is a two-element array containing a lower bound and an upper bound for the estimate. The estimate includes the size of all the items in the table, plus the size of all attributes projected into all of the local secondary indexes on that table. Use this estimate to measure whether a local secondary index is approaching its size limit. The estimate is subject to change over time; therefore, do not rely on the precision or accuracy of the estimate.

pirsConsumedCapacity - The capacity units consumed by the PutItem operation. The data returned includes the total provisioned throughput consumed, along with statistics for the table and any indexes involved in the operation. ConsumedCapacity is only returned if the ReturnConsumedCapacity parameter was specified. For more information, see Provisioned Throughput in the Amazon DynamoDB Developer Guide .

pirsAttributes - The attribute values as they appeared before the PutItem operation, but only if ReturnValues is specified as ALL_OLD in the request. Each element consists of an attribute name and an attribute value.

pirsResponseStatus - -- | The response status code.

data PutItemResponse

Represents the output of a PutItem operation.

See: putItemResponse smart constructor.

Response Lenses

pirsItemCollectionMetrics :: Lens' PutItemResponse (Maybe ItemCollectionMetrics)

Information about item collections, if any, that were affected by the PutItem operation. ItemCollectionMetrics is only returned if the ReturnItemCollectionMetrics parameter was specified. If the table does not have any local secondary indexes, this information is not returned in the response. Each ItemCollectionMetrics element consists of: * ItemCollectionKey - The partition key value of the item collection. This is the same as the partition key value of the item itself. * SizeEstimateRange - An estimate of item collection size, in gigabytes. This value is a two-element array containing a lower bound and an upper bound for the estimate. The estimate includes the size of all the items in the table, plus the size of all attributes projected into all of the local secondary indexes on that table. Use this estimate to measure whether a local secondary index is approaching its size limit. The estimate is subject to change over time; therefore, do not rely on the precision or accuracy of the estimate.

pirsConsumedCapacity :: Lens' PutItemResponse (Maybe ConsumedCapacity)

The capacity units consumed by the PutItem operation. The data returned includes the total provisioned throughput consumed, along with statistics for the table and any indexes involved in the operation. ConsumedCapacity is only returned if the ReturnConsumedCapacity parameter was specified. For more information, see Provisioned Throughput in the Amazon DynamoDB Developer Guide .

pirsAttributes :: Lens' PutItemResponse (HashMap Text AttributeValue)

The attribute values as they appeared before the PutItem operation, but only if ReturnValues is specified as ALL_OLD in the request. Each element consists of an attribute name and an attribute value.

pirsResponseStatus :: Lens' PutItemResponse Int

- | The response status code.