Network.AWS.DynamoDB.UpdateTable

Description

Modifies the provisioned throughput settings, global secondary indexes, or DynamoDB Streams settings for a given table.

You can only perform one of the following operations at once:

Modify the provisioned throughput settings of the table.

Enable or disable Streams on the table.

Remove a global secondary index from the table.

Create a new global secondary index on the table. Once the index begins backfilling, you can use UpdateTable to perform other operations.

UpdateTable is an asynchronous operation; while it is executing, the table status changes from ACTIVE to UPDATING . While it is UPDATING , you cannot issue another UpdateTable request. When the table returns to the ACTIVE state, the UpdateTable operation is complete.

Creating a Request

updateTable

Arguments

:: Text
utTableName

-> UpdateTable

Creates a value of UpdateTable with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

utAttributeDefinitions - An array of attributes that describe the key schema for the table and indexes. If you are adding a new global secondary index to the table, AttributeDefinitions must include the key element(s) of the new index.

utProvisionedThroughput - The new provisioned throughput settings for the specified table or index.

utGlobalSecondaryIndexUpdates - An array of one or more global secondary indexes for the table. For each index in the array, you can request one action: * Create - add a new global secondary index to the table. * Update - modify the provisioned throughput settings of an existing global secondary index. * Delete - remove a global secondary index from the table. For more information, see Managing Global Secondary Indexes in the Amazon DynamoDB Developer Guide .

utStreamSpecification - Represents the DynamoDB Streams configuration for the table.

utTableName - The name of the table to be updated.

data UpdateTable

Represents the input of an UpdateTable operation.

See: updateTable smart constructor.

Request Lenses

utAttributeDefinitions :: Lens' UpdateTable [AttributeDefinition]

An array of attributes that describe the key schema for the table and indexes. If you are adding a new global secondary index to the table, AttributeDefinitions must include the key element(s) of the new index.

utProvisionedThroughput :: Lens' UpdateTable (Maybe ProvisionedThroughput)

The new provisioned throughput settings for the specified table or index.

utGlobalSecondaryIndexUpdates :: Lens' UpdateTable [GlobalSecondaryIndexUpdate]

An array of one or more global secondary indexes for the table. For each index in the array, you can request one action: * Create - add a new global secondary index to the table. * Update - modify the provisioned throughput settings of an existing global secondary index. * Delete - remove a global secondary index from the table. For more information, see Managing Global Secondary Indexes in the Amazon DynamoDB Developer Guide .

utStreamSpecification :: Lens' UpdateTable (Maybe StreamSpecification)

Represents the DynamoDB Streams configuration for the table.

utTableName :: Lens' UpdateTable Text

The name of the table to be updated.

Destructuring the Response

updateTableResponse

Arguments

:: Int
utrsResponseStatus

-> UpdateTableResponse

Creates a value of UpdateTableResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

utrsTableDescription - Represents the properties of the table.

utrsResponseStatus - -- | The response status code.

data UpdateTableResponse

Represents the output of an UpdateTable operation.

See: updateTableResponse smart constructor.

Response Lenses

utrsTableDescription :: Lens' UpdateTableResponse (Maybe TableDescription)

Represents the properties of the table.

utrsResponseStatus :: Lens' UpdateTableResponse Int

- | The response status code.