Network.AWS.DynamoDB.UpdateTable

Description

Updates the provisioned throughput for the given table. Setting the throughput for a table helps you manage performance and is part of the provisioned throughput feature of DynamoDB. The provisioned throughput values can be upgraded or downgraded based on the maximums and minimums listed in the Limits section in the Amazon DynamoDB Developer Guide. The table must be in the ACTIVE state for this operation to succeed. UpdateTable is an asynchronous operation; while executing the operation, the table is in the UPDATING state. While the table is in the UPDATING state, the table still has the provisioned throughput from before the call. The new provisioned throughput setting is in effect only when the table returns to the ACTIVE state after the UpdateTable operation. You cannot add, modify or delete indexes using UpdateTable. Indexes can only be defined at table creation time. Modify Provisioned Write Throughput This example changes both the provisioned read and write throughput of the Thread table to 10 capacity units. { TableDescription: { AttributeDefinitions: [ { AttributeName: ForumName, AttributeType: S }, { AttributeName: LastPostDateTime, AttributeType: S }, { AttributeName: Subject, AttributeType: S } ], CreationDateTime: 1.363801528686E9, ItemCount: 0, KeySchema: [ { AttributeName: ForumName, KeyType: HASH }, { AttributeName: Subject, KeyType: RANGE } ], LocalSecondaryIndexes: [ { IndexName: LastPostIndex, IndexSizeBytes: 0, ItemCount: 0, KeySchema: [ { AttributeName: ForumName, KeyType: HASH }, { AttributeName: LastPostDateTime, KeyType: RANGE } ], Projection: { ProjectionType: KEYS_ONLY } } ], ProvisionedThroughput: { LastIncreaseDateTime: 1.363801701282E9, NumberOfDecreasesToday: 0, ReadCapacityUnits: 5, WriteCapacityUnits: 5 }, TableName: Thread, TableSizeBytes: 0, TableStatus: UPDATING } }.

Request

data UpdateTable

Represents the input of an UpdateTable operation.

Request constructor

updateTable

Arguments

:: Text
utTableName

-> UpdateTable

Smart constructor for the minimum required parameters to construct a valid UpdateTable request.

The fields accessible through corresponding lenses are:

TableName :: Text

ProvisionedThroughput :: Maybe ProvisionedThroughput

GlobalSecondaryIndexUpdates :: [GlobalSecondaryIndexUpdate]

Request lenses

utTableName :: Lens' UpdateTable Text

The name of the table to be updated.

utProvisionedThroughput :: Lens' UpdateTable (Maybe ProvisionedThroughput)

Represents the provisioned throughput settings for a specified table or index. The settings can be modified using the UpdateTable operation. For current minimum and maximum provisioned throughput values, see Limits in the Amazon DynamoDB Developer Guide.

utGlobalSecondaryIndexUpdates :: Lens' UpdateTable [GlobalSecondaryIndexUpdate]

An array of one or more global secondary indexes on the table, together with provisioned throughput settings for each index.

Response

data UpdateTableResponse

Represents the output of an UpdateTable operation.

Response constructor

updateTableResponse :: UpdateTableResponse

Smart constructor for the minimum required parameters to construct a valid UpdateTableResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

TableDescription :: Maybe TableDescription

Response lenses

utrTableDescription :: Lens' UpdateTableResponse (Maybe TableDescription)

Represents the properties of a table.