Network.AWS.DynamoDB.UpdateTable

Description

Updates the provisioned throughput for the given table, or manages the global secondary indexes on the table.

You can increase or decrease the table's provisioned throughput values within the maximums and minimums listed in the Limits section in the AmazonDynamoDB Developer Guide.

In addition, you can use UpdateTable to add, modify or delete global secondary indexes on the table. For more information, see Managing GlobalSecondary Indexes in the Amazon DynamoDB Developer Guide.

The table must be in the ACTIVE state for UpdateTable to succeed. UpdateTable is an asynchronous operation; while executing the operation, the table is in the UPDATING state. While the table is in the UPDATING state, the table still has the provisioned throughput from before the call. The table's new provisioned throughput settings go into effect when the table returns to the ACTIVE state; at that point, the UpdateTable operation is complete.

http://docs.aws.amazon.com/amazondynamodb/latest/APIReference/API_UpdateTable.html

Request

data UpdateTable

Request constructor

updateTable

Arguments

:: Text
utTableName

-> UpdateTable

UpdateTable constructor.

The fields accessible through corresponding lenses are:

utAttributeDefinitions :: [AttributeDefinition]

utGlobalSecondaryIndexUpdates :: [GlobalSecondaryIndexUpdate]

utProvisionedThroughput :: Maybe ProvisionedThroughput

utTableName :: Text

Request lenses

utAttributeDefinitions :: Lens' UpdateTable [AttributeDefinition]

An array of attributes that describe the key schema for the table and indexes. If you are adding a new global secondary index to the table, AttributeDefinitions must include the key element(s) of the new index.

utGlobalSecondaryIndexUpdates :: Lens' UpdateTable [GlobalSecondaryIndexUpdate]

An array of one or more global secondary indexes for the table. For each index in the array, you can request one action:

Create - add a new global secondary index to the table.

Update - modify the provisioned throughput settings of an existing global secondary index.

Delete - remove a global secondary index from the table.

utProvisionedThroughput :: Lens' UpdateTable (Maybe ProvisionedThroughput)

utTableName :: Lens' UpdateTable Text

The name of the table to be updated.

Response

data UpdateTableResponse

Response constructor

updateTableResponse :: UpdateTableResponse

UpdateTableResponse constructor.

The fields accessible through corresponding lenses are:

utrTableDescription :: Maybe TableDescription

Response lenses

utrTableDescription :: Lens' UpdateTableResponse (Maybe TableDescription)