Network.AWS.DynamoDB.V2012_08_10.DeleteTable

Description

The DeleteTable operation deletes a table and all of its items. After a DeleteTable request, the specified table is in the DELETING state until DynamoDB completes the deletion. If the table is in the ACTIVE state, you can delete it. If a table is in CREATING or UPDATING states, then DynamoDB returns a ResourceInUseException. If the specified table does not exist, DynamoDB returns a ResourceNotFoundException. If table is already in the DELETING state, no error is returned. DynamoDB might continue to accept data read and write operations, such as GetItem and PutItem, on a table in the DELETING state until the table deletion is complete. When you delete a table, any indexes on that table are also deleted. Use the DescribeTable API to check the status of the table. Delete a Table This example deletes the Reply table. { TableDescription: { ItemCount: 0, ProvisionedThroughput: { NumberOfDecreasesToday: 0, ReadCapacityUnits: 5, WriteCapacityUnits: 5 }, TableName: Reply, TableSizeBytes: 0, TableStatus: DELETING } }.

Documentation

data DeleteTable

Constructors

DeleteTable

Fields

_dtiTableName :: Text

The name of the table to delete.

dtiTableName :: Iso' DeleteTable Text

data DeleteTableResponse

Constructors

DeleteTableResponse

Fields

_dtoTableDescription :: Maybe TableDescription

Represents the properties of a table.

dtoTableDescription :: Iso' DeleteTableResponse (Maybe TableDescription)