Network.AWS.DynamoDB.V2012_08_10.ListTables

Description

Returns an array of all the tables associated with the current account and endpoint. List Tables This example requests a list of tables, starting with a table named comp2 and ending after three table names have been returned. { LastEvaluatedTableName: Thread, TableNames: [Forum,Reply,Thread] }.

Request

data ListTables

Represents the input of a ListTables operation.

Request constructor

mkListTables :: ListTables

Smart constructor for the minimum required parameters to construct a valid ListTables request.

Request lenses

ltExclusiveStartTableName :: Lens' ListTables (Maybe Text)

The name of the table that starts the list. If you already ran a ListTables operation and received a LastEvaluatedTableName value in the response, use that value here to continue the list.

ltLimit :: Lens' ListTables (Maybe Integer)

A maximum number of table names to return.

Response

data ListTablesResponse

Represents the output of a ListTables operation.

Response constructor

mkListTablesResponse :: ListTablesResponse

Smart constructor for the minimum required parameters to construct a valid ListTablesResponse response.

This constructor is provided for convenience and testing purposes.

Response lenses

ltrTableNames :: Lens' ListTablesResponse [Text]

The names of the tables associated with the current account at the current endpoint.

ltrLastEvaluatedTableName :: Lens' ListTablesResponse (Maybe Text)

The name of the last table in the current list, only if some tables for the account and endpoint have not been returned. This value does not exist in a response if all table names are already returned. Use this value as the ExclusiveStartTableName in a new request to continue the list until all the table names are returned.