Network.AWS.DynamoDB.V2012_08_10.Types

Description

Amazon DynamoDB is a fully managed NoSQL database service that provides fast and predictable performance with seamless scalability. You can use Amazon DynamoDB to create a database table that can store and retrieve any amount of data, and serve any level of request traffic. Amazon DynamoDB automatically spreads the data and traffic for the table over a sufficient number of servers to handle the request capacity specified by the customer and the amount of data stored, while maintaining consistent and fast performance.

Service

data DynamoDB

Supported version (2012-08-10) of the Amazon DynamoDB service.

AttributeAction

data AttributeAction

Constructors

AttributeActionAdd
ADD

AttributeActionDelete
DELETE

AttributeActionPut
PUT

ComparisonOperator

data ComparisonOperator

Constructors

ComparisonOperatorBeginsWith
BEGINS_WITH

ComparisonOperatorBetween
BETWEEN

ComparisonOperatorContains
CONTAINS

ComparisonOperatorEq
EQ

ComparisonOperatorGe
GE

ComparisonOperatorGt
GT

ComparisonOperatorIn
IN

ComparisonOperatorLe
LE

ComparisonOperatorLt
LT

ComparisonOperatorNe
NE

ComparisonOperatorNotContains
NOT_CONTAINS

ComparisonOperatorNotNull
NOT_NULL

ComparisonOperatorNull
NULL

ConditionalOperator

data ConditionalOperator

Constructors

ConditionalOperatorAnd
AND

ConditionalOperatorOr
OR

IndexStatus

data IndexStatus

Constructors

IndexStatusActive
ACTIVE

IndexStatusCreating
CREATING

IndexStatusDeleting
DELETING

IndexStatusUpdating
UPDATING

KeyType

data KeyType

Constructors

KeyTypeHash
HASH

KeyTypeRange
RANGE

ProjectionType

data ProjectionType

Constructors

ProjectionTypeAll
ALL

ProjectionTypeInclude
INCLUDE

ProjectionTypeKeysOnly
KEYS_ONLY

ReturnConsumedCapacity

data ReturnConsumedCapacity

Constructors

ReturnConsumedCapacityIndexes
INDEXES

ReturnConsumedCapacityNone
NONE

ReturnConsumedCapacityTotal
TOTAL

ReturnItemCollectionMetrics

data ReturnItemCollectionMetrics

Constructors

ReturnItemCollectionMetricsNone
NONE

ReturnItemCollectionMetricsSize
SIZE

ReturnValue

data ReturnValue

Constructors

ReturnValueAllNew
ALL_NEW

ReturnValueAllOld
ALL_OLD

ReturnValueNone
NONE

ReturnValueUpdatedNew
UPDATED_NEW

ReturnValueUpdatedOld
UPDATED_OLD

ScalarAttributeType

data ScalarAttributeType

Constructors

ScalarAttributeTypeB
B

ScalarAttributeTypeN
N

ScalarAttributeTypeS
S

Select

data Select

Constructors

SelectAllAttributes
ALL_ATTRIBUTES

SelectAllProjectedAttributes
ALL_PROJECTED_ATTRIBUTES

SelectCount
COUNT

SelectSpecificAttributes
SPECIFIC_ATTRIBUTES

TableStatus

data TableStatus

Constructors

TableStatusActive
ACTIVE

TableStatusCreating
CREATING

TableStatusDeleting
DELETING

TableStatusUpdating
UPDATING

Capacity

data Capacity

The amount of throughput consumed on the table affected by the operation.

mkCapacity :: Capacity

Smart constructor for the minimum required fields to construct a valid Capacity data type to populate a request.

cCapacityUnits :: Lens' Capacity (Maybe Double)

The total number of capacity units consumed on a table or an index.

DeleteRequest

data DeleteRequest

A request to perform a DeleteItem operation.

mkDeleteRequest

Arguments

:: Map Text AttributeValue
drKey

-> DeleteRequest

Smart constructor for the minimum required fields to construct a valid DeleteRequest data type to populate a request.

drKey :: Lens' DeleteRequest (Map Text AttributeValue)

A map of attribute name to attribute values, representing the primary key of the item to delete. All of the table's primary key attributes must be specified, and their data types must match those of the table's key schema.

GlobalSecondaryIndexUpdate

data GlobalSecondaryIndexUpdate

Represents the new provisioned throughput settings to apply to a global secondary index.

mkGlobalSecondaryIndexUpdate :: GlobalSecondaryIndexUpdate

Smart constructor for the minimum required fields to construct a valid GlobalSecondaryIndexUpdate data type to populate a request.

gsiuUpdate :: Lens' GlobalSecondaryIndexUpdate (Maybe UpdateGlobalSecondaryIndexAction)

The name of a global secondary index, along with the updated provisioned throughput settings that are to be applied to that index.

PutRequest

data PutRequest

A request to perform a PutItem operation.

mkPutRequest

Arguments

:: Map Text AttributeValue
prItem

-> PutRequest

Smart constructor for the minimum required fields to construct a valid PutRequest data type to populate a request.

prItem :: Lens' PutRequest (Map Text AttributeValue)

A map of attribute name to attribute values, representing the primary key of an item to be processed by PutItem. All of the table's primary key attributes must be specified, and their data types must match those of the table's key schema. If any attributes are present in the item which are part of an index key schema for the table, their types must match the index key schema.

AttributeDefinition

data AttributeDefinition

Represents an attribute for describing the key schema for the table and indexes.

mkAttributeDefinition

Arguments

:: Text
adAttributeName

-> ScalarAttributeType
adAttributeType

-> AttributeDefinition

Smart constructor for the minimum required fields to construct a valid AttributeDefinition data type to populate a request.

adAttributeName :: Lens' AttributeDefinition Text

A name for the attribute.

adAttributeType :: Lens' AttributeDefinition ScalarAttributeType

The data type for the attribute.

AttributeValue

data AttributeValue

Represents the data for an attribute. You can set one, and only one, of the elements.

mkAttributeValue :: AttributeValue

Smart constructor for the minimum required fields to construct a valid AttributeValue data type to populate a request.

avS :: Lens' AttributeValue (Maybe Text)

A String data type.

avN :: Lens' AttributeValue (Maybe Text)

A Number data type.

avB :: Lens' AttributeValue (Maybe Base64)

A Binary data type.

avSS :: Lens' AttributeValue [Text]

A String set data type.

avNS :: Lens' AttributeValue [Text]

Number set data type.

avBS :: Lens' AttributeValue [Base64]

A Binary set data type.

AttributeValueUpdate

data AttributeValueUpdate

For the UpdateItem operation, represents the attributes to be modified,the action to perform on each, and the new value for each. You cannot use UpdateItem to update any primary key attributes. Instead, you will need to delete the item, and then use PutItem to create a new item with new attributes. Attribute values cannot be null; string and binary type attributes must have lengths greater than zero; and set type attributes must not be empty. Requests with empty values will be rejected with a ValidationException.

mkAttributeValueUpdate :: AttributeValueUpdate

Smart constructor for the minimum required fields to construct a valid AttributeValueUpdate data type to populate a request.

avuValue :: Lens' AttributeValueUpdate (Maybe AttributeValue)

Represents the data for an attribute. You can set one, and only one, of the elements.

avuAction :: Lens' AttributeValueUpdate (Maybe AttributeAction)

Specifies how to perform the update. Valid values are PUT, DELETE, and ADD. The behavior depends on whether the specified primary key already exists in the table. If an item with the specified Key is found in the table: PUT - Adds the specified attribute to the item. If the attribute already exists, it is replaced by the new value. DELETE - If no value is specified, the attribute and its value are removed from the item. The data type of the specified value must match the existing value's data type. If a set of values is specified, then those values are subtracted from the old set. For example, if the attribute value was the set [a,b,c] and the DELETE action specified [a,c], then the final attribute value would be [b]. Specifying an empty set is an error. ADD - If the attribute does not already exist, then the attribute and its values are added to the item. If the attribute does exist, then the behavior of ADD depends on the data type of the attribute: If the existing attribute is a number, and if Value is also a number, then the Value is mathematically added to the existing attribute. If Value is a negative number, then it is subtracted from the existing attribute. If you use ADD to increment or decrement a number value for an item that doesn't exist before the update, DynamoDB uses 0 as the initial value. In addition, if you use ADD to update an existing item, and intend to increment or decrement an attribute value which does not yet exist, DynamoDB uses 0 as the initial value. For example, suppose that the item you want to update does not yet have an attribute named itemcount, but you decide to ADD the number 3 to this attribute anyway, even though it currently does not exist. DynamoDB will create the itemcount attribute, set its initial value to 0, and finally add 3 to it. The result will be a new itemcount attribute in the item, with a value of 3. If the existing data type is a set, and if the Value is also a set, then the Value is added to the existing set. (This is a set operation, not mathematical addition.) For example, if the attribute value was the set [1,2], and the ADD action specified [3], then the final attribute value would be [1,2,3]. An error occurs if an Add action is specified for a set attribute and the attribute type specified does not match the existing set type. Both sets must have the same primitive data type. For example, if the existing data type is a set of strings, the Value must also be a set of strings. The same holds true for number sets and binary sets. This action is only valid for an existing attribute whose data type is number or is a set. Do not use ADD for any other data types. If no item with the specified Key is found: PUT - DynamoDB creates a new item with the specified primary key, and then adds the attribute. DELETE - Nothing happens; there is no attribute to delete. ADD - DynamoDB creates an item with the supplied primary key and number (or set of numbers) for the attribute value. The only data types allowed are number and number set; no other data types can be specified.

Condition

data Condition

Represents a selection criteria for a Query or Scan operation. For a Query operation, the condition specifies the key attributes to use when querying a table or an index. For a Scan operation, the condition is used to evaluate the scan results and return only the desired values. Multiple conditions are ANDed together. In other words, all of the conditions must be met to be included in the output.

mkCondition

Arguments

:: ComparisonOperator
c1ComparisonOperator

-> Condition

Smart constructor for the minimum required fields to construct a valid Condition data type to populate a request.

c1AttributeValueList :: Lens' Condition [AttributeValue]

One or more values to evaluate against the supplied attribute. This list contains exactly one value, except for a BETWEEN or IN comparison, in which case the list contains two values. For type Number, value comparisons are numeric. String value comparisons for greater than, equals, or less than are based on ASCII character code values. For example, a is greater than A, and aa is greater than B. For a list of code values, see http:/en.wikipedia.orgwiki/ASCII#ASCII_printable_characters. For Binary, DynamoDB treats each byte of the binary data as unsigned when it compares binary values, for example when evaluating query expressions.

c1ComparisonOperator :: Lens' Condition ComparisonOperator

A comparator for evaluating attributes. For example, equals, greater than, less than, etc. Valid comparison operators for Query: EQ | LE | LT | GE | GT | BEGINS_WITH | BETWEEN Valid comparison operators for Scan: EQ | NE | LE | LT | GE | GT | NOT_NULL | NULL | CONTAINS | NOT_CONTAINS | BEGINS_WITH | IN | BETWEEN For information on specifying data types in JSON, see JSON Data Format in the Amazon DynamoDB Developer Guide. The following are descriptions of each comparison operator. EQ : Equal. AttributeValueList can contain only one AttributeValue of type String, Number, or Binary (not a set). If an item contains an AttributeValue of a different type than the one specified in the request, the value does not match. For example, {S:"6"} does not equal {N:"6"}. Also, {N:"6"} does not equal {NS:["6", "2", "1"]}. NE : Not equal. AttributeValueList can contain only one AttributeValue of type String, Number, or Binary (not a set). If an item contains an AttributeValue of a different type than the one specified in the request, the value does not match. For example, {S:"6"} does not equal {N:"6"}. Also, {N:"6"} does not equal {NS:["6", "2", "1"]}. LE : Less than or equal. AttributeValueList can contain only one AttributeValue of type String, Number, or Binary (not a set). If an item contains an AttributeValue of a different type than the one specified in the request, the value does not match. For example, {S:"6"} does not equal {N:"6"}. Also, {N:"6"} does not compare to {NS:["6", "2", "1"]}. LT : Less than. AttributeValueList can contain only one AttributeValue of type String, Number, or Binary (not a set). If an item contains an AttributeValue of a different type than the one specified in the request, the value does not match. For example, {S:"6"} does not equal {N:"6"}. Also, {N:"6"} does not compare to {NS:["6", "2", "1"]}. GE : Greater than or equal. AttributeValueList can contain only one AttributeValue of type String, Number, or Binary (not a set). If an item contains an AttributeValue of a different type than the one specified in the request, the value does not match. For example, {S:"6"} does not equal {N:"6"}. Also, {N:"6"} does not compare to {NS:["6", "2", "1"]}. GT : Greater than. AttributeValueList can contain only one AttributeValue of type String, Number, or Binary (not a set). If an item contains an AttributeValue of a different type than the one specified in the request, the value does not match. For example, {S:"6"} does not equal {N:"6"}. Also, {N:"6"} does not compare to {NS:["6", "2", "1"]}. NOT_NULL : The attribute exists. NULL : The attribute does not exist. CONTAINS : checks for a subsequence, or value in a set. AttributeValueList can contain only one AttributeValue of type String, Number, or Binary (not a set). If the target attribute of the comparison is a String, then the operation checks for a substring match. If the target attribute of the comparison is Binary, then the operation looks for a subsequence of the target that matches the input. If the target attribute of the comparison is a set (SS, NS, or BS), then the operation checks for a member of the set (not as a substring). NOT_CONTAINS : checks for absence of a subsequence, or absence of a value in a set. AttributeValueList can contain only one AttributeValue of type String, Number, or Binary (not a set). If the target attribute of the comparison is a String, then the operation checks for the absence of a substring match. If the target attribute of the comparison is Binary, then the operation checks for the absence of a subsequence of the target that matches the input. If the target attribute of the comparison is a set (SS, NS, or BS), then the operation checks for the absence of a member of the set (not as a substring). BEGINS_WITH : checks for a prefix. AttributeValueList can contain only one AttributeValue of type String or Binary (not a Number or a set). The target attribute of the comparison must be a String or Binary (not a Number or a set). IN : checks for exact matches. AttributeValueList can contain more than one AttributeValue of type String, Number, or Binary (not a set). The target attribute of the comparison must be of the same type and exact value to match. A String never matches a String set. BETWEEN : Greater than or equal to the first value, and less than or equal to the second value. AttributeValueList must contain two AttributeValue elements of the same type, either String, Number, or Binary (not a set). A target attribute matches if the target value is greater than, or equal to, the first element and less than, or equal to, the second element. If an item contains an AttributeValue of a different type than the one specified in the request, the value does not match. For example, {S:"6"} does not compare to {N:"6"}. Also, {N:"6"} does not compare to {NS:["6", "2", "1"]}.

ConsumedCapacity

data ConsumedCapacity

Represents the capacity units consumed by an operation. The data returned includes the total provisioned throughput consumed, along with statistics for the table and any indexes involved in the operation. ConsumedCapacity is only returned if it was asked for in the request. For more information, see Provisioned Throughput in the Amazon DynamoDB Developer Guide.

mkConsumedCapacity :: ConsumedCapacity

Smart constructor for the minimum required fields to construct a valid ConsumedCapacity data type.

ConsumedCapacity is exclusively used in responses and this constructor is provided for convenience and testing purposes.

ccTableName :: Lens' ConsumedCapacity (Maybe Text)

The name of the table that was affected by the operation.

ccCapacityUnits :: Lens' ConsumedCapacity (Maybe Double)

The total number of capacity units consumed by the operation.

ccTable :: Lens' ConsumedCapacity (Maybe Capacity)

The amount of throughput consumed on the table affected by the operation.

ccLocalSecondaryIndexes :: Lens' ConsumedCapacity (Map Text Capacity)

The amount of throughput consumed on each local index affected by the operation.

ccGlobalSecondaryIndexes :: Lens' ConsumedCapacity (Map Text Capacity)

The amount of throughput consumed on each global index affected by the operation.

ExpectedAttributeValue

data ExpectedAttributeValue

Represents an attribute value used with conditional DeleteItem, PutItem or UpdateItem operations. DynamoDB will check to see if the attribute value already exists; or if the attribute exists and has a particular value before updating it.

mkExpectedAttributeValue :: ExpectedAttributeValue

Smart constructor for the minimum required fields to construct a valid ExpectedAttributeValue data type to populate a request.

eavValue :: Lens' ExpectedAttributeValue (Maybe AttributeValue)

Represents the data for an attribute. You can set one, and only one, of the elements.

eavExists :: Lens' ExpectedAttributeValue (Maybe Bool)

Causes DynamoDB to evaluate the value before attempting a conditional operation: If Exists is true, DynamoDB will check to see if that attribute value already exists in the table. If it is found, then the operation succeeds. If it is not found, the operation fails with a ConditionalCheckFailedException. If Exists is false, DynamoDB assumes that the attribute value does not exist in the table. If in fact the value does not exist, then the assumption is valid and the operation succeeds. If the value is found, despite the assumption that it does not exist, the operation fails with a ConditionalCheckFailedException. The default setting for Exists is true. If you supply a Value all by itself, DynamoDB assumes the attribute exists: You don't have to set Exists to true, because it is implied. DynamoDB returns a ValidationException if: Exists is true but there is no Value to check. (You expect a value to exist, but don't specify what that value is.) Exists is false but you also specify a Value. (You cannot expect an attribute to have a value, while also expecting it not to exist.) If you specify more than one condition for Exists, then all of the conditions must evaluate to true. (In other words, the conditions are ANDed together.) Otherwise, the conditional operation will fail.

eavComparisonOperator :: Lens' ExpectedAttributeValue (Maybe ComparisonOperator)

eavAttributeValueList :: Lens' ExpectedAttributeValue [AttributeValue]

GlobalSecondaryIndex

data GlobalSecondaryIndex

Represents a global secondary index.

mkGlobalSecondaryIndex

Arguments

:: Text
gsiIndexName

-> List1 KeySchemaElement
gsiKeySchema

-> Projection
gsiProjection

-> ProvisionedThroughput
gsiProvisionedThroughput

-> GlobalSecondaryIndex

Smart constructor for the minimum required fields to construct a valid GlobalSecondaryIndex data type to populate a request.

gsiIndexName :: Lens' GlobalSecondaryIndex Text

The name of the global secondary index. The name must be unique among all other indexes on this table.

gsiKeySchema :: Lens' GlobalSecondaryIndex (List1 KeySchemaElement)

The complete key schema for a global secondary index, which consists of one or more pairs of attribute names and key types (HASH or RANGE).

gsiProjection :: Lens' GlobalSecondaryIndex Projection

Represents attributes that are copied (projected) from the table into an index. These are in addition to the primary key attributes and index key attributes, which are automatically projected.

gsiProvisionedThroughput :: Lens' GlobalSecondaryIndex ProvisionedThroughput

Represents the provisioned throughput settings for a specified table or index. The settings can be modified using the UpdateTable operation. For current minimum and maximum provisioned throughput values, see Limits in the Amazon DynamoDB Developer Guide.

GlobalSecondaryIndexDescription

data GlobalSecondaryIndexDescription

Represents the properties of a global secondary index.

mkGlobalSecondaryIndexDescription :: GlobalSecondaryIndexDescription

Smart constructor for the minimum required fields to construct a valid GlobalSecondaryIndexDescription data type.

GlobalSecondaryIndexDescription is exclusively used in responses and this constructor is provided for convenience and testing purposes.

gsidIndexName :: Lens' GlobalSecondaryIndexDescription (Maybe Text)

The name of the global secondary index.

gsidKeySchema :: Lens' GlobalSecondaryIndexDescription (Maybe (List1 KeySchemaElement))

The complete key schema for the global secondary index, consisting of one or more pairs of attribute names and key types (HASH or RANGE).

gsidProjection :: Lens' GlobalSecondaryIndexDescription (Maybe Projection)

Represents attributes that are copied (projected) from the table into an index. These are in addition to the primary key attributes and index key attributes, which are automatically projected.

gsidIndexStatus :: Lens' GlobalSecondaryIndexDescription (Maybe IndexStatus)

The current state of the global secondary index: CREATING - The index is being created, as the result of a CreateTable or UpdateTable operation. UPDATING - The index is being updated, as the result of a CreateTable or UpdateTable operation. DELETING - The index is being deleted, as the result of a DeleteTable operation. ACTIVE - The index is ready for use.

gsidProvisionedThroughput :: Lens' GlobalSecondaryIndexDescription (Maybe ProvisionedThroughputDescription)

Represents the provisioned throughput settings for the table, consisting of read and write capacity units, along with data about increases and decreases.

gsidIndexSizeBytes :: Lens' GlobalSecondaryIndexDescription (Maybe Integer)

The total size of the specified index, in bytes. DynamoDB updates this value approximately every six hours. Recent changes might not be reflected in this value.

gsidItemCount :: Lens' GlobalSecondaryIndexDescription (Maybe Integer)

The number of items in the specified index. DynamoDB updates this value approximately every six hours. Recent changes might not be reflected in this value.

ItemCollectionMetrics

data ItemCollectionMetrics

Information about item collections, if any, that were affected by the operation. ItemCollectionMetrics is only returned if it was asked for in the request. If the table does not have any local secondary indexes, this information is not returned in the response.

mkItemCollectionMetrics :: ItemCollectionMetrics

Smart constructor for the minimum required fields to construct a valid ItemCollectionMetrics data type to populate a request.

icmItemCollectionKey :: Lens' ItemCollectionMetrics (Map Text AttributeValue)

The hash key value of the item collection. This is the same as the hash key of the item.

icmSizeEstimateRangeGB :: Lens' ItemCollectionMetrics [Double]

An estimate of item collection size, measured in gigabytes. This is a two-element array containing a lower bound and an upper bound for the estimate. The estimate includes the size of all the items in the table, plus the size of all attributes projected into all of the local secondary indexes on that table. Use this estimate to measure whether a local secondary index is approaching its size limit. The estimate is subject to change over time; therefore, do not rely on the precision or accuracy of the estimate.

KeySchemaElement

data KeySchemaElement

Represents a single element of a key schema. A key schema specifies the attributes that make up the primary key of a table, or the key attributes of an index.

mkKeySchemaElement

Arguments

:: Text
kseAttributeName

-> KeyType
kseKeyType

-> KeySchemaElement

Smart constructor for the minimum required fields to construct a valid KeySchemaElement data type to populate a request.

kseAttributeName :: Lens' KeySchemaElement Text

The name of a key attribute.

kseKeyType :: Lens' KeySchemaElement KeyType

The attribute data, consisting of the data type and the attribute value itself.

KeysAndAttributes

data KeysAndAttributes

Represents a set of primary keys and, for each key, the attributes to retrieve from the table.

mkKeysAndAttributes

Arguments

:: List1 (Map Text AttributeValue)
kaaKeys

-> KeysAndAttributes

Smart constructor for the minimum required fields to construct a valid KeysAndAttributes data type to populate a request.

kaaKeys :: Lens' KeysAndAttributes (List1 (Map Text AttributeValue))

The primary key attribute values that define the items and the attributes associated with the items.

kaaAttributesToGet :: Lens' KeysAndAttributes (Maybe (List1 Text))

One or more attributes to retrieve from the table or index. If no attribute names are specified then all attributes will be returned. If any of the specified attributes are not found, they will not appear in the result.

kaaConsistentRead :: Lens' KeysAndAttributes (Maybe Bool)

The consistency of a read operation. If set to true, then a strongly consistent read is used; otherwise, an eventually consistent read is used.

LocalSecondaryIndex

data LocalSecondaryIndex

Represents a local secondary index.

mkLocalSecondaryIndex

Arguments

:: Text
lsiIndexName

-> List1 KeySchemaElement
lsiKeySchema

-> Projection
lsiProjection

-> LocalSecondaryIndex

Smart constructor for the minimum required fields to construct a valid LocalSecondaryIndex data type to populate a request.

lsiIndexName :: Lens' LocalSecondaryIndex Text

The name of the local secondary index. The name must be unique among all other indexes on this table.

lsiKeySchema :: Lens' LocalSecondaryIndex (List1 KeySchemaElement)

The complete key schema for the local secondary index, consisting of one or more pairs of attribute names and key types (HASH or RANGE).

lsiProjection :: Lens' LocalSecondaryIndex Projection

Represents attributes that are copied (projected) from the table into an index. These are in addition to the primary key attributes and index key attributes, which are automatically projected.

LocalSecondaryIndexDescription

data LocalSecondaryIndexDescription

Represents the properties of a local secondary index.

mkLocalSecondaryIndexDescription :: LocalSecondaryIndexDescription

Smart constructor for the minimum required fields to construct a valid LocalSecondaryIndexDescription data type.

LocalSecondaryIndexDescription is exclusively used in responses and this constructor is provided for convenience and testing purposes.

lsidIndexName :: Lens' LocalSecondaryIndexDescription (Maybe Text)

Represents the name of the local secondary index.

lsidKeySchema :: Lens' LocalSecondaryIndexDescription (Maybe (List1 KeySchemaElement))

The complete index key schema, which consists of one or more pairs of attribute names and key types (HASH or RANGE).

lsidProjection :: Lens' LocalSecondaryIndexDescription (Maybe Projection)

Represents attributes that are copied (projected) from the table into an index. These are in addition to the primary key attributes and index key attributes, which are automatically projected.

lsidIndexSizeBytes :: Lens' LocalSecondaryIndexDescription (Maybe Integer)

The total size of the specified index, in bytes. DynamoDB updates this value approximately every six hours. Recent changes might not be reflected in this value.

lsidItemCount :: Lens' LocalSecondaryIndexDescription (Maybe Integer)

The number of items in the specified index. DynamoDB updates this value approximately every six hours. Recent changes might not be reflected in this value.

Projection

data Projection

Represents attributes that are copied (projected) from the table into an index. These are in addition to the primary key attributes and index key attributes, which are automatically projected.

mkProjection :: Projection

Smart constructor for the minimum required fields to construct a valid Projection data type to populate a request.

pProjectionType :: Lens' Projection (Maybe ProjectionType)

The set of attributes that are projected into the index: KEYS_ONLY - Only the index and primary keys are projected into the index. INCLUDE - Only the specified table attributes are projected into the index. The list of projected attributes are in NonKeyAttributes. ALL - All of the table attributes are projected into the index.

pNonKeyAttributes :: Lens' Projection (Maybe (List1 Text))

Represents the non-key attribute names which will be projected into the index. For local secondary indexes, the total count of NonKeyAttributes summed across all of the local secondary indexes, must not exceed 20. If you project the same attribute into two different indexes, this counts as two distinct attributes when determining the total.

ProvisionedThroughput

data ProvisionedThroughput

Represents the provisioned throughput settings for a specified table or index. The settings can be modified using the UpdateTable operation. For current minimum and maximum provisioned throughput values, see Limits in the Amazon DynamoDB Developer Guide.

mkProvisionedThroughput

Arguments

:: Integer
ptReadCapacityUnits

-> Integer
ptWriteCapacityUnits

-> ProvisionedThroughput

Smart constructor for the minimum required fields to construct a valid ProvisionedThroughput data type to populate a request.

ptReadCapacityUnits :: Lens' ProvisionedThroughput Integer

The maximum number of strongly consistent reads consumed per second before DynamoDB returns a ThrottlingException. For more information, see Specifying Read and Write Requirements in the Amazon DynamoDB Developer Guide.

ptWriteCapacityUnits :: Lens' ProvisionedThroughput Integer

The maximum number of writes consumed per second before DynamoDB returns a ThrottlingException. For more information, see Specifying Read and Write Requirements in the Amazon DynamoDB Developer Guide.

ProvisionedThroughputDescription

data ProvisionedThroughputDescription

The provisioned throughput settings for the table, consisting of read and write capacity units, along with data about increases and decreases.

mkProvisionedThroughputDescription :: ProvisionedThroughputDescription

Smart constructor for the minimum required fields to construct a valid ProvisionedThroughputDescription data type to populate a request.

ptdLastIncreaseDateTime :: Lens' ProvisionedThroughputDescription (Maybe ISO8601)

The date and time of the last provisioned throughput increase for this table.

ptdLastDecreaseDateTime :: Lens' ProvisionedThroughputDescription (Maybe ISO8601)

The date and time of the last provisioned throughput decrease for this table.

ptdNumberOfDecreasesToday :: Lens' ProvisionedThroughputDescription (Maybe Integer)

The number of provisioned throughput decreases for this table during this UTC calendar day. For current maximums on provisioned throughput decreases, see Limits in the Amazon DynamoDB Developer Guide.

ptdReadCapacityUnits :: Lens' ProvisionedThroughputDescription (Maybe Integer)

The maximum number of strongly consistent reads consumed per second before DynamoDB returns a ThrottlingException. Eventually consistent reads require less effort than strongly consistent reads, so a setting of 50 ReadCapacityUnits per second provides 100 eventually consistent ReadCapacityUnits per second.

ptdWriteCapacityUnits :: Lens' ProvisionedThroughputDescription (Maybe Integer)

The maximum number of writes consumed per second before DynamoDB returns a ThrottlingException.

TableDescription

data TableDescription

Represents the properties of a table.

mkTableDescription :: TableDescription

Smart constructor for the minimum required fields to construct a valid TableDescription data type.

TableDescription is exclusively used in responses and this constructor is provided for convenience and testing purposes.

tdAttributeDefinitions :: Lens' TableDescription [AttributeDefinition]

An array of AttributeDefinition objects. Each of these objects describes one attribute in the table and index key schema. Each AttributeDefinition object in this array is composed of: AttributeName - The name of the attribute. AttributeType - The data type for the attribute.

tdTableName :: Lens' TableDescription (Maybe Text)

The name of the table.

tdKeySchema :: Lens' TableDescription (Maybe (List1 KeySchemaElement))

The primary key structure for the table. Each KeySchemaElement consists of: AttributeName - The name of the attribute. KeyType - The key type for the attribute. Can be either HASH or RANGE. For more information about primary keys, see Primary Key in the Amazon DynamoDB Developer Guide.

tdTableStatus :: Lens' TableDescription (Maybe TableStatus)

The current state of the table: CREATING - The table is being created, as the result of a CreateTable operation. UPDATING - The table is being updated, as the result of an UpdateTable operation. DELETING - The table is being deleted, as the result of a DeleteTable operation. ACTIVE - The table is ready for use.

tdCreationDateTime :: Lens' TableDescription (Maybe ISO8601)

The date and time when the table was created, in UNIX epoch time format.

tdProvisionedThroughput :: Lens' TableDescription (Maybe ProvisionedThroughputDescription)

The provisioned throughput settings for the table, consisting of read and write capacity units, along with data about increases and decreases.

tdTableSizeBytes :: Lens' TableDescription (Maybe Integer)

The total size of the specified table, in bytes. DynamoDB updates this value approximately every six hours. Recent changes might not be reflected in this value.

tdItemCount :: Lens' TableDescription (Maybe Integer)

The number of items in the specified table. DynamoDB updates this value approximately every six hours. Recent changes might not be reflected in this value.

tdLocalSecondaryIndexes :: Lens' TableDescription [LocalSecondaryIndexDescription]

Represents one or more local secondary indexes on the table. Each index is scoped to a given hash key value. Tables with one or more local secondary indexes are subject to an item collection size limit, where the amount of data within a given item collection cannot exceed 10 GB. Each element is composed of: IndexName - The name of the local secondary index. KeySchema - Specifies the complete index key schema. The attribute names in the key schema must be between 1 and 255 characters (inclusive). The key schema must begin with the same hash key attribute as the table. Projection - Specifies attributes that are copied (projected) from the table into the index. These are in addition to the primary key attributes and index key attributes, which are automatically projected. Each attribute specification is composed of: ProjectionType - One of the following: KEYS_ONLY - Only the index and primary keys are projected into the index. INCLUDE - Only the specified table attributes are projected into the index. The list of projected attributes are in NonKeyAttributes. ALL - All of the table attributes are projected into the index. NonKeyAttributes - A list of one or more non-key attribute names that are projected into the secondary index. The total count of attributes specified in NonKeyAttributes, summed across all of the secondary indexes, must not exceed 20. If you project the same attribute into two different indexes, this counts as two distinct attributes when determining the total. IndexSizeBytes - Represents the total size of the index, in bytes. DynamoDB updates this value approximately every six hours. Recent changes might not be reflected in this value. ItemCount - Represents the number of items in the index. DynamoDB updates this value approximately every six hours. Recent changes might not be reflected in this value. If the table is in the DELETING state, no information about indexes will be returned.

tdGlobalSecondaryIndexes :: Lens' TableDescription [GlobalSecondaryIndexDescription]

The global secondary indexes, if any, on the table. Each index is scoped to a given hash key value. Each element is composed of: IndexName - The name of the global secondary index. IndexSizeBytes - The total size of the global secondary index, in bytes. DynamoDB updates this value approximately every six hours. Recent changes might not be reflected in this value. IndexStatus - The current status of the global secondary index: CREATING - The index is being created. UPDATING - The index is being updated. DELETING - The index is being deleted. ACTIVE - The index is ready for use. ItemCount - The number of items in the global secondary index. DynamoDB updates this value approximately every six hours. Recent changes might not be reflected in this value. KeySchema - Specifies the complete index key schema. The attribute names in the key schema must be between 1 and 255 characters (inclusive). The key schema must begin with the same hash key attribute as the table. Projection - Specifies attributes that are copied (projected) from the table into the index. These are in addition to the primary key attributes and index key attributes, which are automatically projected. Each attribute specification is composed of: ProjectionType - One of the following: KEYS_ONLY - Only the index and primary keys are projected into the index. INCLUDE - Only the specified table attributes are projected into the index. The list of projected attributes are in NonKeyAttributes. ALL - All of the table attributes are projected into the index. NonKeyAttributes - A list of one or more non-key attribute names that are projected into the secondary index. The total count of attributes specified in NonKeyAttributes, summed across all of the secondary indexes, must not exceed 20. If you project the same attribute into two different indexes, this counts as two distinct attributes when determining the total. ProvisionedThroughput - The provisioned throughput settings for the global secondary index, consisting of read and write capacity units, along with data about increases and decreases. If the table is in the DELETING state, no information about indexes will be returned.

UpdateGlobalSecondaryIndexAction

data UpdateGlobalSecondaryIndexAction

The name of a global secondary index, along with the updated provisioned throughput settings that are to be applied to that index.

mkUpdateGlobalSecondaryIndexAction

Arguments

:: Text
ugsiaIndexName

-> ProvisionedThroughput
ugsiaProvisionedThroughput

-> UpdateGlobalSecondaryIndexAction

Smart constructor for the minimum required fields to construct a valid UpdateGlobalSecondaryIndexAction data type to populate a request.

ugsiaIndexName :: Lens' UpdateGlobalSecondaryIndexAction Text

The name of the global secondary index to be updated.

ugsiaProvisionedThroughput :: Lens' UpdateGlobalSecondaryIndexAction ProvisionedThroughput

Represents the provisioned throughput settings for a specified table or index. The settings can be modified using the UpdateTable operation. For current minimum and maximum provisioned throughput values, see Limits in the Amazon DynamoDB Developer Guide.

WriteRequest

data WriteRequest

Represents an operation to perform - either DeleteItem or PutItem. You can only specify one of these operations, not both, in a single WriteRequest. If you do need to perform both of these operations, you will need to specify two separate WriteRequest objects.

mkWriteRequest :: WriteRequest

Smart constructor for the minimum required fields to construct a valid WriteRequest data type to populate a request.

wrPutRequest :: Lens' WriteRequest (Maybe PutRequest)

A request to perform a PutItem operation.

wrDeleteRequest :: Lens' WriteRequest (Maybe DeleteRequest)

A request to perform a DeleteItem operation.